Macro rtx_package::DefMacro

macro_rules! DefMacro {
    ($proto:literal, None) => { ... };
    ($prototype:literal, sub [( $($var:ident),* )]
    $body:block $($input:tt)*) => { ... };
    ($proto:expr, sub [$args:ident]
    $body:block $($input:tt)*) => { ... };
    ($proto:expr, $body:block $($input:tt)*) => { ... };
    ($proto:literal, $expansion:literal $($input:tt)*) => { ... };
    ($cs:expr, $parameters:expr, sub [$args:ident]
    $body:block $($input:tt)*) => { ... };
    ($cs:literal, None, $expansion:literal) => { ... };
    ($cs:literal, None, $expansion:literal, $($input:tt)*) => { ... };
    ($cs:literal, None, $expansion:expr) => { ... };
    ($cs:expr, None, $expansion:literal) => { ... };
    ($cs:expr, None, $body:block) => { ... };
    ($cs:expr, None, $expansion:expr) => { ... };
    ($cs:expr, None, $expansion:literal, $($input:tt)+) => { ... };
    ($cs:expr, None, $expansion:expr, $($input:tt)+) => { ... };
    ($cs:expr, $parameters:expr, $expansion:expr) => { ... };
    ($proto:literal, None $($input:tt)*) => { ... };
    ($cs:expr, $replacement:expr, $expansion:expr, $($input:tt)*) => { ... };
}

A prototype will be parsed into a command sequence and a list of parameters. Any macro arguments will be substituted for parameter indicators (eg #1) in the Tokens or tokenized string and the result is used as the expansion of the control sequence. If a closure is used, it is called at expansion time and should return a list of tokens as its result.