Macro rtx_package::DefRegister

macro_rules! DefRegister {
    ($proto:expr => $value:expr) => { ... };
    ($proto:expr, $value:expr) => { ... };
    ($cs:expr, None, $value:expr) => { ... };
    ($cs:expr, None, $value:expr, $($input:tt)+) => { ... };
    ($proto:expr, $value:expr, $($input:tt)+) => { ... };
    ($proto:expr => $value:expr, $($input:tt)+) => { ... };
}

Defines a register with value as the initial value

(a Number, Dimension, Glue, MuGlue or Tokens — I haven’t handled Box’s yet). Usually, the prototype is just the control sequence, but registers are also handled by prototypes like \count{Number}. DefRegister arranges that the register value can be accessed when a numeric, dimension, … value is being read, and also defines the control sequence for assignment.