Module rtx_package::package

Re-exports

pub use log;

pub use self::latex_functions::*;

pub use self::tex_functions::*;

Modules

alltt_sty

amsmath_sty

amsthm_sty

article_cls

aux_macros

Auxiliary macros

binding

The programmable API foundation for creating bindings of LaTeX sty/cls libraries

cite_sty

comment

TeX comments as standalone objects

comment_sty

common

Common abstractions that are useful at various stages of Core processing

definition

All possible definitions for TeX-native commands (expandable, primitive, constructor,…)

digested

a shared interface for digested objects

document

An abstraction layer over the converted XML document

etex

etoolbox_sty

font

fontenc_sty

gullet

The Gullet is responsible for reading Tokens and other data from the Mouth

hyperref_sty

ieeetran_cls

inputenc_sty

keyval

A LaTeX-like Key-Value object

keyvals

A collection of Key-Value objects, typically from a single LaTeX argument

latex

latex_functions

latexml_sty

ligature

Rules for combining together characters and other text rules

list

A list of Digested objects

mouth

The mouth is a thin interface over a file, responsible for reading characters and associating them with catcodes

multido_sty

parameter

The abstraction layer used by the Gullet to read arguments for the various kinds of TeX object definitions

pathname

helper methods for file system paths

pdftex

radix

“radix” may be a misnomer here. Primarily used to generate labels, or uniquifying suffixes to make ID’s simple radix conversion utilities

rewrite

Rules for rewriting the constructed XML document, after core processing has completed

setup_binding_language

state

A global, singleton, mutable State - hosts almost all TeX-facing runtime information for the conversion

stomach

The stomach is an abstraction responsible for digesting Tokens and Registers prepared by the Gullet into Boxes

t1_fontmap

t1 font encoding

t1enc_def

t1enc_sty

tbox

A TeX-like digested Box

tex

tex_functions

textcase_sty

textcomp_sty

token

The Token and its Catcode

tokens

Grouping Tokens together

url_sty

utf8_def

util

Auxilary utilities that do not participate in the main conversion abstraction

verbatim_sty

whatsit

A TeX-like digested Whatsit

Macros

AssignRegister

Debug

Error

Explode

Explode a string into a list of tokens, all w/catcode OTHER (except space).

ExplodeText

Similar to Explode, but convert letters to catcode LETTER and others to OTHER Hopefully, this is essentially correct WITHOUT resorting to catcode lookup?

Fatal

Font

build a Font from key=>val pairs

FontDirective

build a FontDirective from key=>val pairs (currently only FontDirective::Asset is supported in this macro)

Info

NewDefault

given a struct $name, create a new instance of it using the given key=>val pairs and complete the remaining entries via the Default instance

NewDefaultV

Just like NewDefault, but adds a mandatory .into_option() to all values

Number

Roman

Convert the number to upper case roman numerals, returning a list of LaTeXML::Core::Token

SetCounter

T_ACTIVE

macro for an ACTIVE char token

T_ALIGN

macro for an ALIGN “&” token

T_ARG

macro for a numbered ARG token

T_BEGIN

macro for a BEGIN “{” token

T_COMMENT

macro for a COMMENT content token

T_CR

macro for a CR “\n” token

T_CS

macro for a command sequence token

T_END

macro for an END “}” token

T_LETTER

macro for a LETTER token

T_MARKER

macro for a tracing MARKER token

T_MATH

macro for a MATH “$” token

T_OTHER

macro for an OTHER code token

T_PARAM

macro for a PARAM “#” token

T_SMUGGLE_THE

macro for a SmuggleThe token (see gullet::invoke_and_read_x_token)

T_SPACE

macro for a SPACE “ “ token

T_SUB

macro for a SUB “_” token

T_SUPER

macro for a SUPER “^” token

Token

Token constructor macro (defaults to OTHER code)

Tokens

convenience macro for assembling a Tokens object from different pieces (Token, Vec<Token>, Tokens)

Warn

after_digest

after_digest_simple

after_digest_single

before_digest

before_digest_simple

before_digest_single

construct

count_unpack_opt

count_unpack_ref

count_unpack_to_string

fatal

fontmap

Instantiates a Font, using the Font fields as keys, and calling .into() for each value. The specification can be partial - missing fields are taken via the Default trait.

forbidMath

generate_message

getter

A closure for obtaining a RegisterValue, usually owned by a Register getter.

lazy_static

map

A flexary macro for constructing HashMap<String, T> maps, where T is generic.

mixvec

A variant on vec! where each argument receives an additional .into() call best used with an outer context that explicitly provides the expected type, such as

noreplacement

primitiveproc

create a PrimitiveClosure from the pieces, with a forced empty return value

print_stderr

prints a to STDERR without a line break

println_stderr

prints a single line to STDERR

prop_bool

prop_digested

prop_str

prop_string

prop_whatsit

properties

raw_char_map

A flexary macro for constructing HashMap<char, T> maps, where T is generic

raw_map

A flexary macro for constructing HashMap<K, V> maps, where K and V are both generic (inferred at time of use)

reader

reader_predigest

replacement

requireMath

reversion

reversion_digested

rewrite_replace_sub

roman

Convert the number to lower case roman numerals, returning a list of LaTeXML::Core::Token

s

The s! macro is a briefer alias for format!, otherwise completely identical

set

Simple generic helper for HashSet<K,V> creation Source: https://riptutorial.com/rust/example/4149/create-a-hashset-macro

setter

sizersub

static_map

! Helper macros for quicker and more idiomatic construction and access to data structures. A flexary macro for constructing HashMap<&'static str, &'static str> maps

stored_map

A flexary macro for constructing HashMap<String, Stored> maps

string_map

A flexary macro for constructing HashMap<String, String> maps

tagsub

transfer_opt_default

Assumption: $receiver is HashMap<String,String>. If, and only if, $has_receiver does not have a value at slot $val, and $struct_source has a set value at $val, then transfers (with ownership) the $val field of a $struct_source into the $receiver.

undigested

unpack

unpack_opt

An alternative to unpack!(args=>arg1,arg2...) which allows for optional arguments.

unpack_opt_ref

Meant to be used for unpacking &Vec<Option> in particular.

unpack_to_number

unpack_to_string

unpacks a Vec<ArgWrap> argument into a flexary number of String-typed variables uses the usual .to_string() method from the Display trait.

unpack_to_token

unref

Convenience macro to flexibly unpack a collection of Vec<ArgWrap> arguments into individual Tokens variables.

Structs

ARG_HOLE

regex for the usual argument placeholders #1-#9

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

BeginItemizeOptions

configuration for begin_itemize

CONDITIONAL_CS_RE

regex for the prefix of a conditional command sequence

Conditional

A Conditional definition; Expandable.

ConditionalOptions

configurations for a conditional.

ConstructorOptions

configuration for creating a new Constructor

Core

The Core conversion runtime

CoreOptions

Configuration for the Core processing

DIRTY_ID_IDIOM_RE

regex for ${}^{label}$

Digested

An Arc-guarded abstraction for any object encountered at the “digested” phase of processing

Dimension

Document

Error

Expandable

ExpandableOptions

FindFileOptions

configuration for searching for a file in the local filesystem

Float

Font

This struct is a little interesting, as we want to pass overrides that partially modify (via a merge) the current font, in each definitional binding. To accommodate that with this struct, every single field needs to be an Option, in order to unambiguously tell the “intend” of override (Some) vs no intent (None).

Glue

Gullet

HAS_ARG_OR_CS

regex for a TeX argument specifier or any command sequence

InputDefinitionOptions

a configuration for loading LaTeX definition files (such as .sty, .cls, and their bindings)

InputOptions

configuration for input of a TeX source (content files mostly)

KeyVals

KeyValsOptions

LEADING_PROTOCOL_RE

regex for the prefix of a protocol (such as literal:)

Ligature

List

Lists can contain any Digested items, such as boxes, whatsits or other lists

Locator

MathLigatureOptions

MathPrimitive

MathPrimitiveOptions

Mouth

MouthOptions

MuDimension

MuGlue

NON_ID_CHARSET_RE

regex for characters not expected in a usual id attribute

Namespace

An xml namespace

NewCounterOptions

! ! # Counters ! ! This is modelled on LaTeX’s counter mechanisms, but since it also ! provides support for ID’s, even where there is no visible reference number, ! it is defined in general. ! These id’s should be both unique, and parallel the visible reference numbers ! (as much as possible). Also, for consistency, we add id’s to unnumbered ! document elements (eg from \section*); this requires an additional counter ! (eg. UNsection) and mechanisms to track it. configuration for new_counter

Node

An xml node

Number

Parameter

Parameters

Primitive

PrimitiveOptions

Regex

A compiled regular expression for matching Unicode strings.

Register

A struct representing a TeX register

RegisterOptions

configuration for creating a new Register

RequireOptions

Resource

Rewrite

RewriteOptions

RwLock

A reader-writer lock

SPACES_RE

regex for one-or-more spaces

State

The State efficiently maintain the bindings in a TeX-like fashion. bindings associate data with keys (eg definitions with macro names) and respect TeX grouping; that is, an assignment is only in effect until the current group (opened by \bgroup) is closed (by \egroup).

Stomach

The Stomach is responsible for digesting tokens into boxes, lists, etc.

TILDE_NOISE_RE

regex for a strange noisy TeX \\~{}

TRAILING_SLASH_RE

regex for a trailing slash (trivial, but aids replacement of said slash)

TagOptions

Tbox

Box is a Rust keyword, so we use “Tbox” instead, as in “TeX Box”

Token

The core immutable syntactic primitive resulting from TeX’s read-in and expansion process We allow the fields to be public, so that we can use builder macros such as

Tokens

Tokens are a thin wrapper over a vector of Token objects usually read from a Mouth. They are usually treated as an immutable interface, an have to be consumed via .unlist() for access to the underlying data.

VecDeque

A double-ended queue implemented with a growable ring buffer.

Whatsit

Represents a digested object that can generate arbitrary elements in the XML Document.

Enums

ArgWrap

Catcode

A Token category code, as in TeX

ConditionalType

! Conditionals Control sequence definitions. ! These represent the control sequences for conditionals, as well as ! \else, \or and \fi. classify the standard pieces of a conditional

Cow

A clone-on-write smart pointer.

DigestedData

These are all kinds of data which we consider officially supported as outputs from the digestion phase of TeX, i.e. from invoking a token.

ErrorCategory

ErrorTarget

ExpansionBody

FontDirective

RegisterType

The type of a TeX Register

RegisterValue

The values that can be read by, and stored in, a Register

Reversion

Scope

installation scope in the State tables

Stored

The original global State (in Perl) allowed arbitrary values. To stay consistent, we create an extremely permissive struct that affords all essential kinds of values that appear essential.

TagOptionName

TexMode

The current TeX processing mode

Constants

MOCK_TOKEN

T_EMPTY

T_RELAX

XML_NS

Traits

BoxOps

Common operations for Box-like (digested) data

Definition

IntoBoolResult

Creates Result<bool> from some type T

IntoDigestedOptionResult

IntoDigestedResult

Creates a Result<Vec<Digested>> from some type T

IntoFontField

IntoOption

A trait for auto-wrapping a generic type T into Option, where Y can be inferred from context. (useful in macro helpers, such as NewDefaultV!)

IntoPropertiesResult

IntoRegisterValueOption

Creates an Option<RegisterValue> from some type T. Useful for Register getter closures

IntoResultArgWrap

Create a Result<ArgWrap> from any concrete type that Gullet may have a reader for. Used in auto-casting the data fetched by Parameter readers

IntoTokensResult

A trait for creating Result<Tokens> from all sensible concrete types one could return from e.g. a DefMacro closure

NumericOps

Object

Base object for all LaTeXML Objects.

Functions

add_to_counter

increments a named counter by a Number

begin_itemize

Prepare for an list (itemize/enumerate/description/etc) by determining the right counter (level) and binding the right \item ( $type@item, if $type is defined)

build_invocation

clean_bib_key

cleans a string down to characters acceptable for a bibliography key

clean_id

cleans a string down to characters acceptable for an id attribute

clean_label

cleans a string down to characters acceptable for a label attribute

clean_url

cleans a string down to characters acceptable for a URL

coerce_cs

unconditionally wraps a CS token around a string

compose_url

builds a complete url from fragments

convert_latex_args

Convert a LaTeX-style argument spec to our Package form. Ie. given $nargs and $optional, being the two optional arguments to something like \newcommand, convert it to the form we use

counter_value

Fetches the value associated with the counter C<$ctr>.

def_conditional

Define a conditional control sequence. Its processing takes place in the Gullet. The test is applied to the arguments (if any), which determines which branch is executed. If the test is undefined, the conditional is a “user defined” one; Two additional primitives are defined \footrue and \foofalse; the test is then determined by the most recently called of those.

def_constructor

The Constructor is where LaTeXML really starts getting interesting; invoking the control sequence will generate an arbitrary XML fragment in the document tree. More specifically: during digestion, the arguments will be read and digested, creating a Whatsit to represent the object. During absorption by the Document, the Whatsit will generate the XML fragment according to the compiled_replacement.

def_environment

Defines an Environment that generates a specific XML fragment. compiled_replacement is of the same form as for DefConstructor, but will generally include reference to the #body property. Upon encountering a \begin{env}: the mode is switched, if needed, else a new group is opened; then the environment name is noted; the beforeDigest hook is run. Then the Whatsit representing the begin command (but ultimately the whole environment) is created and the after_digest_begin hook is run. Next, the body will be digested and collected until the balancing \end{env}. Then, any after_digest hook is run, the environment is ended, finally the mode is ended or the group is closed. The body and \end{env} whatsit are added to the \begin{env}’s whatsit as body and trailer, respectively.

def_macro

Defines the macro expansion for a cs: a macro control sequence that reads parameters as specified by paramlist and is expanded during macro expansion time in the Gullet. See ExpansionBody for the possible kinds of expansion material.

def_math

Define a Mathematical symbol or function. There are two sets of cases: (1) If the presentation appears to be TeX code, we create an XMDual, since the presentation may end up with structure, etc. (2) But if the presentation is a simple string, or unicode, it is just the content of the symbol; even if the function takes arguments.

def_math_constructor

Uses of DefMath without arguments, but with constructor-like options, are realized via a Constructor definition

def_math_dual

Advanced math replacements require a XMDual representation

def_math_primitive

EXPERIMENT: Introduce an intermediate case for simple symbols Define a primitive that will create a Tbox with the appropriate set of XMTok attributes.

def_primitive

Defines a primitive control sequence; a primitive is processed during digestion (in the Stomach), after macro expansion but before Construction time. Primitive control sequences generate Boxes or Lists, generally containing basic Unicode content, rather than structured XML. Primitive control sequences are also executed for side effect during digestion, effecting changes to the State.

def_register

Defines a register with value as the initial value (a Number, Dimension, Glue, MuGlue or Tokens — I haven’t handled Box’s yet). Usually, the prototype is just the control sequence, but registers are also handled by prototypes like \count{Number}. DefRegister arranges that the register value can be accessed when a numeric, dimension, … value is being read, and also defines the control sequence for assignment.

digest_if

digest_literal

digest_text

dispatch

do_expand

dualize_arglist

Given a list of Tokens (to be expanded into mathematical objects) return two lists: (1) The Tokens’ wrapped in an XMAarg, with an ID added (2) a corresponding list of Tokens creating XMRef’s to those IDs

find_file

search for a file as prescribed by a FindFileOptions configuration

floatformat

Utility for formatting sane numbers.

input

This is essentially the \input equivalent; we are most likely expecting to get actual content, (possibly with definitions included, as well) but might actually be getting pure definitions, (like a proper style file) in which case we may really want to load a binding. Note that generic style files (non-latex) often have a .tex extension.

input_content

used for cases when the file (or data) is plain TeX material that is expected to contribute content to the document (as opposed to pure definitions). A Mouth is opened onto the file, and subsequent reading and/or digestion will pull Tokens from that Mouth until it is exhausted, or closed.

input_definitions

TODO: Flesh out with the full infrastructure, incremental functionality for now.

install_tag

is_definable

Check if the token is not yet defined, or let to \relax

is_defined

Is defined in the LaTeX-y sense of also not being let to \relax.

is_defined_token

Token variant of is_defined. Defined in the LaTeX-y sense of also not being let to \relax.

load_binding

loads a binding from the main binding dispatcher, if available+found

load_class

load_external_binding

loads a binding from an external binding dispatcher, if available+found

load_font_map

load_tex_content

maybe_note_label

Use to note a discovered label to support label-derived refererence numbers Can by used by \label, among others. Note we only record the label if it hasn’t already been peeked, and consumed.

merge_font

new_counter

Defines a new counter named $ctr. If within is defined, ctr will be reset whenever within is incremented.

note_begin

note_end

note_progress

note_progress_detailed

preload_font_map

process_options

ref_step_counter

Analog of \refstepcounter, steps the counter and returns a hash containing the keys refnum and id. This makes it suitable for use in a properties option to constructors. The id is generated in parallel with the reference number to assist debugging.

ref_step_id

For UN-numbered units. Like RefStepCounter, but only steps the “uncounter”, and returns only the id; This is useful for unnumbered cases of objects that normally get both a refnum and id.

ref_step_item_counter

Create id, and tags for an itemize type \item

require_package

This (and FindFile) needs to evolve a bit to support reading raw .sty (.def, etc) files from the standard texmf directories. Maybe even use kpsewhich itself (INSTEAD of pathname_find ???) Another potentially useful option might be that if we are reading a raw file, perhaps it should just get digested immediately, since it shouldn’t contribute any boxes.

require_resource

rescue_caption_counters

Copies the current id, tags, and inlist counter values into whatsit properties

reset_counter

Resets the counter ctr to zero.

resource_type

roman_aux

auxiliary helper for roman

select_relaxng_schema

Selects the RelaxNG schema defining the XML output language

step_counter

Analog of \stepcounter, steps the counter and returns the expansion of \the$ctr Usually you should use ref_step_counter(ctr) instead.

step_counter_gullet

DG: This is a dirty refactor that avoids the Stomach threading for DefMath — is there a conceptually better organization?

Type Definitions

BeforeDigestClosure

ConditionalClosure

ConstructionClosure

DigestedReversionClosure

DigestionClosure

ExpansionClosure

FontClosure

FontTestClosure

HashMap

Type alias for a hashmap using the fx hash algorithm.

LigatureMatcher

PrimitiveClosure

PrimitiveFn

ReaderClosure

ReplacementClosure

Result

ReversionClosure