Module rtx_package::package::common

Common abstractions that are useful at various stages of Core processing

Modules

arena

An arena for interning strings (global, mutable, single thread only)

cleaners

def_parser

dimension

error

float

font

glue

ligature

local_assignments

This module can benefit from some thinking over and refactoring in principle the “local” scope variables in Perl are a completely standalone feature (it’s global namespace value shadowing until scope expiration) … so the push/pop can be modeled with sentry values getting created and dropped (TODO)…

locator

model

mudimension

muglue

number

numeric_ops

object

relaxng

Extract Model information from a RelaxNG schema

store

xml

Structs

Config

Enums

DataSize

DigestionMode

InputFormat

OutputFormat

Type Aliases

BindingDispatcher