Module rtx_package::package::common::arena

An arena for interning strings (global, mutable, single thread only)

Note: works under the assumption of a single threaded, short-lived process, where memory starvation will not be an issue. It would need a hard reset if the same process does multiple conversions, and a different implementation if one needs a thread-local arena.

Modules

data

Structs

SymHashMap

A convenience abstraction over a String-keyed HashMap typically used for HashMap<String,Stored> states. The goal is to support both a string interface, as well as the interned tickets interface, while avoiding String allocations internally.

Statics

ANY_SYM

the unique symbol for str value “ANY”

CAPTURE_SYM

the unique symbol for str value “ltx:Capture”

DONT_EXPAND_SYM

the unique symbol for str value “\dont_expand”

DTD_SYM

the unique symbol for str value “DTD”

EMPTY_SYM

the unique symbol for the empty str value “”

FONT_SYM

the unique symbol for str value “font”

GLOBAL_DEFS_SYM

the unique symbol for str value “\globaldefs”

H_COMMENT_SYM

the unique symbol for str value “#COMMENT”

H_DEFAULT_SYM

the unique symbol for str value “#default”

H_DOC_SYM

the unique symbol for str value “#Document”

H_DTD_SYM

the unique symbol for str value “#DTD”

H_PCDATA_SYM

the unique symbol for str value “#PCDATA”

H_PI_SYM

the unique symbol for str value “#ProcessingInstruction”

LTX_P_SYM

the unique symbol for str value “ltx:p”

LTX_STAR_SYM

the unique symbol for str value “ltx:*”

MATH_SYM

the unique symbol for str value “math”

RELAXNG_SYM

the unique symbol for str value “RelaxNG”

TEXT_SYM

the unique symbol for str value “text”

WILD_CARD_SYM

the unique symbol for str value “WildCard”

XML_ID_SYM

the unique symbol for str value “xml:id”

Functions

into_pin

join

len

pin

Assign a string into the arena, returning a unique symbol associated with it

pin_char

pin_static

Assign a static str into the arena, returning a unique symbol associated with it

to_string

with

with2

with3

with_many

Type Aliases

SymStr