Module rtx_package::package::common::model

Structs

Model

TagFrame

Constants

LTX_NAMESPACE

Statics

MODEL

Functions

add_schema_declaration

add_tag_attribute

add_tag_content

can_contain

Can an element with (qualified name) tag contain a child element?

can_contain_sym

can_have_attribute

decode_qname

Given a Qualified name, possibly prefixed with a namespace prefix, as defined by the code namespace mapping, return the NamespaceURI and localname.

get_document_namespace

get_document_namespace_prefix

get_namespace

get_namespace_prefix

In the following: $forattribute is 1 if the namespace is for an attribute (in which case, there must be a non-empty prefix) $probe, if non 0, just test for namespace, without creating an entry if missing. Get the (code) prefix associated with $namespace, creating a dummy prefix and signalling an error if none has been registered.

get_node_document_qname

Same as get_node_qname, but using the Document namespace prefixes

get_node_qname

Get the node’s qualified name in standard form Ie. using the registered (code) prefix for that namespace. NOTE: Reconsider how Capture & WildCard should be integrated!?!

get_tag_contents

get_tags

initialize_model

is_in_schema_class

is_node_in_schema_class

is_permissive

load_schema

register_document_namespace

register_namespace

set_doc_type

set_model

set_relaxng_schema

set_schema

set_schema_class

with_code_namespaces

with_node_qname

with_schema_data

Type Definitions

IndirectModel