Struct rtx_package::package::common::model::Model

pub struct Model {
    pub schema: Option<Relaxng>,
    pub schema_data: Option<Vec<String, Global>>,
    pub schema_class: HashMap<String, HashSet<String, RandomState>, RandomState>,
    pub code_namespace_prefixes: HashMap<String, String, RandomState>,
    pub code_namespaces: HashMap<String, String, RandomState>,
    pub document_namespace_prefixes: HashMap<String, String, RandomState>,
    pub document_namespaces: HashMap<String, String, RandomState>,
    pub permissive: bool,
    pub no_compiled: bool,
    pub debug_mode: bool,
    pub namespace_errors: u8,
    pub tagprop: HashMap<String, TagFrame, RandomState>,
}

Fields
schema: Option<Relaxng>schema_data: Option<Vec<String, Global>>schema_class: HashMap<String, HashSet<String, RandomState>, RandomState>code_namespace_prefixes: HashMap<String, String, RandomState>code_namespaces: HashMap<String, String, RandomState>document_namespace_prefixes: HashMap<String, String, RandomState>document_namespaces: HashMap<String, String, RandomState>permissive: boolno_compiled: booldebug_mode: boolnamespace_errors: u8tagprop: HashMap<String, TagFrame, RandomState>
Implementations

impl Model

pub fn new() -> Model

pub fn set_doc_type(&mut self, roottag: String, publicid: String, systemid: String)

pub fn set_relaxng_schema(&mut self, schema: String)

pub fn add_schema_declaration(&self, document: &mut Document)

pub fn load_schema(&mut self, search_paths: &[&str]) -> &Option<Relaxng>

pub fn describe_model(&self)

pub fn get_xpath<'o>(&'o self, document: &'o Document) -> XPath

pub fn register_namespace(
&mut self,
codeprefix: &str,
namespace_opt: Option<String>
)

Namespaces

There are TWO namespace mappings!!! One for coding, one for the DocType.

Coding: this namespace mapping associates prefixes to namespace URIs for use in the latexml code, constructors and such. This must be a one to one mapping and there are no default namespaces. Document: this namespace mapping associates prefixes to namespace URIs as used in the generated document, and will be the set of prefixes used in the generated output. This mapping may also use a prefix of “#default” which is for the unprefixed form of elements (not used for attributes!)

pub fn register_document_namespace(
&mut self,
docprefix: &str,
namespace_opt: Option<String>
)

pub fn get_document_namespace_prefix(
&mut self,
namespace: &str,
forattribute: bool,
probe: bool
) -> Option<String>

pub fn get_document_namespace(
&mut self,
docprefix: &str,
probe: bool
) -> Option<String>

pub fn get_namespace_prefix(
&mut self,
namespace: &str,
_forattribute: bool,
probe: bool
) -> Option<String>

In the following: $forattribute is 1 if the namespace is for an attribute (in which case, there must be a non-empty prefix) $probe, if non 0, just test for namespace, without creating an entry if missing. Get the (code) prefix associated with $namespace, creating a dummy prefix and signalling an error if none has been registered.

pub fn get_namespace(&mut self, codeprefix: &str, probe: bool) -> Option<String>

pub fn get_node_qname(&self, node: &Node) -> String

Get the node’s qualified name in standard form Ie. using the registered (code) prefix for that namespace. NOTE: Reconsider how Capture & WildCard should be integrated!?!

pub fn get_node_document_qname(&mut self, node: &Node) -> String

Same as get_node_qname, but using the Document namespace prefixes

pub fn decode_qname(&mut self, codetag: &str) -> (Option<String>, String)

Given a Qualified name, possibly prefixed with a namespace prefix, as defined by the code namespace mapping, return the NamespaceURI and localname.

pub fn can_contain(&mut self, tag: &str, child: &str) -> bool

Can an element with (qualified name) tag contain a child element?

pub fn can_have_attribute(&mut self, tag: &str, attrib: &str) -> bool

pub fn is_node_in_schema_class(&self, class_name: &str, tag: &Node) -> bool

pub fn is_in_schema_class(&self, class_name: &str, tag: &str) -> bool

pub fn load_compiled_schema(&mut self, path: &str)

TODO: This is another component that would fit perfectly as a compiler plugin. For now, simply reimplementing the runtime loading of LaTeXML.model as-is from Model.pm

pub fn get_tags(&self) -> Vec<String, Global>

pub fn get_tag_contents(&self, tag: &str) -> Vec<&str, Global>

pub fn add_tag_content(&mut self, tag: &str, elements: Vec<&str, Global>)

pub fn get_tag_attributes(&self, tag: &str) -> Vec<&str, Global>

pub fn add_tag_attribute(&mut self, tag: &str, attributes: Vec<&str, Global>)

pub fn set_schema_class(
&mut self,
classname: &str,
content: HashSet<String, RandomState>
)