Struct rtx_package::package::digested::Digested

pub struct Digested(_);

An Rc-guarded abstraction for any object encountered at the “digested” phase of processing

Implementations

impl Digested

pub fn data(&self) -> &DigestedData

immutably borrow the inner Digested data

pub fn value_of(&self) -> i64

Obtain the i64 value of the digested object, iff it wraps a RegisterValue

pub fn pt_value(&self, prec: Option<u8>) -> f64

Obtain the f64 value of the digested object, iff it wraps a RegisterValue

pub fn any<F>(&self, check: F) -> boolwhere F: FnMut(&Digested) -> bool,

Predicate check - true if any element of the current object passes the check

pub fn all<F>(&self, check: F) -> boolwhere F: FnMut(&Digested) -> bool,

Predicate check - true if all elements of the current object passes the check

pub fn is_empty(&self) -> bool

Predicate check - delegates to .is_empty() of the underlying data

pub fn raw_tokens(&self) -> Option<&Tokens>

Provide a way of emulating an Undigested argument, by requesting raw tokens, only when they are preserved – empty otherwise.

pub fn to_attribute(&self) -> String

builds an attribute-friendly String form of the digested object, suitable for XML attributes

pub fn untex(&self, state: &mut State) -> Result<String, Error>

Reverts a digested object to Tokens and extracts a TeX-near string representation of its content

pub fn alignment_cell(&self) -> Option<&RefCell<Alignment>>