Enum rtx_package::package::Stored

pub enum Stored {

    None,
    Bool(bool),
    String(SymbolU32),
    Charcode(u16),
    Int(i64),
    Node(Node),
    Chars(Box<[char]>),
    Fontmap(Rc<[Option<char>]>),
    Strings(Rc<[String]>),
    VecDigested(Vec<Digested>),
    Stash(Vec<(TableName, SymbolU32, Stored)>),
    HashString(HashMap<String, String, BuildHasherDefault<FxHasher>>),
    VecDequeStored(VecDeque<Stored>),
    HashStored(SymHashMap<Stored>),
    HashTagData(HashMap<String, Vec<(String, Option<HashMap<String, String, BuildHasherDefault<FxHasher>>>, Digested)>, BuildHasherDefault<FxHasher>>),
    Catcode(Catcode),
    Token(Token),
    Tokens(Tokens),
    Number(Number),
    Float(Float),
    Glue(Glue),
    MuGlue(MuGlue),
    Dimension(Dimension),
    MuDimension(MuDimension),
    Locator(Box<Locator>),
    Rewrite(Box<Rewrite>),
    Ligature(Box<Ligature>),
    Reversion(Reversion),
    Register(Rc<Register>),
    Expandable(Rc<Expandable>),
    Conditional(Rc<Conditional>),
    Primitive(Rc<Primitive>),
    MathPrimitive(Rc<MathPrimitive>),
    Constructor(Rc<Constructor>),
    Digested(Digested),
    Parameter(Rc<Parameter>),
    Font(Rc<Font>),
    FontDirective(FontDirective),
    Mouth(Rc<RefCell<Mouth>>),
    IfFrame(Rc<RefCell<IfFrame>>),

}

The original global state (in Perl) allowed arbitrary values. To stay consistent, we create an extremely permissive struct that affords all essential kinds of values that appear essential.

Variants

None

if we want to keep a key but make it ‘undef’, set it to None

Bool(bool)

atomic data (Copy)

String(SymbolU32)

atomic data (Clone)

Charcode(u16)

atomic data (Copy)

Int(i64)

atomic data (Copy) note that we currently work with 64-bit integers

Node(Node)

atomic data (Clone)

Chars(Box<[char]>)

boxed char

Fontmap(Rc<[Option<char>]>)

boxed Option

Strings(Rc<[String]>)

boxed collection

VecDigested(Vec<Digested>)

boxed collection (latexml)

Stash(Vec<(TableName, SymbolU32, Stored)>)

the heart of state - a stored Stash table

HashString(HashMap<String, String, BuildHasherDefault<FxHasher>>)

boxed map

VecDequeStored(VecDeque<Stored>)

boxed collection - Stored

HashStored(SymHashMap<Stored>)

boxed map - Stored

HashTagData(HashMap<String, Vec<(String, Option<HashMap<String, String, BuildHasherDefault<FxHasher>>>, Digested)>, BuildHasherDefault<FxHasher>>)

boxed map (latexml)

Catcode(Catcode)

latexml object

Token(Token)

latexml object

Tokens(Tokens)

latexml object

Number(Number)

latexml object

Float(Float)

latexml object

Glue(Glue)

latexml object

MuGlue(MuGlue)

latexml object

Dimension(Dimension)

latexml object

MuDimension(MuDimension)

latexml object

Locator(Box<Locator>)

metadata object

Rewrite(Box<Rewrite>)

latexml object

Ligature(Box<Ligature>)

latexml object

Reversion(Reversion)

latexml object

Register(Rc<Register>)

latexml object (Rc-wrapped)

Expandable(Rc<Expandable>)

latexml object (Rc-wrapped)

Conditional(Rc<Conditional>)

latexml object (Rc-wrapped)

Primitive(Rc<Primitive>)

latexml object (Rc-wrapped)

MathPrimitive(Rc<MathPrimitive>)

latexml object (Rc-wrapped)

Constructor(Rc<Constructor>)

latexml object (Rc-wrapped)

Digested(Digested)

latexml object (Rc-wrapped from within)

Parameter(Rc<Parameter>)

latexml object (Rc-wrapped)

Font(Rc<Font>)

latexml object (Rc-wrapped)

FontDirective(FontDirective)

a stored FontDirective (Font or closure building a Font)

Mouth(Rc<RefCell<Mouth>>)

WALL OF SHAME (interior mutability) – can we dispense with these?

IfFrame(Rc<RefCell<IfFrame>>)

WALL OF SHAME (interior mutability) – can we dispense with these?

Implementations

impl Stored

pub fn cast_to_string_hash( in_map: &SymHashMap<Stored> ) -> HashMap<String, String, BuildHasherDefault<FxHasher>>

helper method that uses ToString::to_string to flatten a map with Stored values

pub fn read_arguments(&self) -> Result<Vec<ArgWrap>, Error>

Dynamic dispatch for Definition’s read_arguments, to circumvent the limitations of using trait objects with Rc<Definition>

pub fn to_attribute(&self) -> String

Uses NumericOps::to_attribute for Stored values supporting it, otherwise ToString::to_string

pub fn to_definition(&self) -> Option<Rc<dyn Definition>>