Function rtx_package::package::def_macro

pub fn def_macro<T>(
    cs: Token,
    paramlist: Option<Parameters>,
    expansion: T,
    options_opt: Option<ExpandableOptions>
) -> Result<(), Error>
where
    T: Into<Option<ExpansionBody>>,

Defines the macro expansion for a cs: a macro control sequence that reads parameters as specified by paramlist and is expanded during macro expansion time in the Gullet. See ExpansionBody for the possible kinds of expansion material.