Struct rtx_package::package::gullet::Gullet

pub struct Gullet {
    pub mouth: Option<MouthRuntime>,
    pub mouthstack: VecDeque<MouthRuntime, Global>,
    pub pending_comments: VecDeque<Token, Global>,
    /* private fields */
}

Fields
mouth: Option<MouthRuntime>mouthstack: VecDeque<MouthRuntime, Global>pending_comments: VecDeque<Token, Global>
Implementations

impl Gullet

pub fn initialize(&mut self)

Initialize (or reset, if reentrant) a Gullet to its default empty state

pub fn get_locator(&self) -> Option<Cow<'_, Locator>>

User feedback for where something (error?) occurred.

pub fn get_location(&self) -> String

pub fn flush_mouth(&mut self)

This flushes a mouth so that it will be automatically closed, next time it’s read Corresponds to TeX’s \endinput

pub fn open_mouth(&mut self, mouth: Mouth, autoclose: bool)

pub fn close_mouth(&mut self, forced: bool) -> Result<(), Error>

pub fn get_mouth(&self) -> Option<&Mouth>

pub fn get_mouth_mut(&mut self) -> Option<&mut Mouth>

pub fn mouth_is_open(&self, mouth: &Mouth) -> bool

pub fn unread(&mut self, tokens: Tokens)

Push the tokens back into the input stream to be re-read.

pub fn unread_mut(&mut self, tokens: &mut Tokens)

same as unread, but drains the tokens from its contents

pub fn unread_one(&mut self, token: Token)

same as unread, but only for a single Token

pub fn unread_vec(&mut self, tokens: Vec<Token, Global>)

same as unread, but does not require the Tokens wrapper