Struct rtx_package::package::model::Model

pub struct Model {
    pub schema: Option<Relaxng>,
    pub schema_data: Option<Vec<SymbolU32, Global>>,
    pub schema_class: HashMap<SymbolU32, HashSet<SymbolU32, BuildHasherDefault<FxHasher>>, BuildHasherDefault<FxHasher>>,
    pub code_namespace_prefixes: HashMap<SymbolU32, SymbolU32, BuildHasherDefault<FxHasher>>,
    pub code_namespaces: HashMap<SymbolU32, SymbolU32, BuildHasherDefault<FxHasher>>,
    pub document_namespace_prefixes: HashMap<SymbolU32, SymbolU32, BuildHasherDefault<FxHasher>>,
    pub document_namespaces: HashMap<SymbolU32, SymbolU32, BuildHasherDefault<FxHasher>>,
    pub permissive: bool,
    pub no_compiled: bool,
    pub debug_mode: bool,
    pub namespace_errors: u8,
    pub tagprop: HashMap<SymbolU32, TagFrame, BuildHasherDefault<FxHasher>>,
}

Fields
schema: Option<Relaxng>schema_data: Option<Vec<SymbolU32, Global>>schema_class: HashMap<SymbolU32, HashSet<SymbolU32, BuildHasherDefault<FxHasher>>, BuildHasherDefault<FxHasher>>code_namespace_prefixes: HashMap<SymbolU32, SymbolU32, BuildHasherDefault<FxHasher>>code_namespaces: HashMap<SymbolU32, SymbolU32, BuildHasherDefault<FxHasher>>document_namespace_prefixes: HashMap<SymbolU32, SymbolU32, BuildHasherDefault<FxHasher>>document_namespaces: HashMap<SymbolU32, SymbolU32, BuildHasherDefault<FxHasher>>permissive: boolno_compiled: booldebug_mode: boolnamespace_errors: u8tagprop: HashMap<SymbolU32, TagFrame, BuildHasherDefault<FxHasher>>
Implementations

impl Model

pub fn new() -> Model

pub fn register_namespace( &mut self, codeprefix: &str, namespace_opt: Option<&str> )

Namespaces

There are TWO namespace mappings!!! One for coding, one for the DocType.

Coding: this namespace mapping associates prefixes to namespace URIs for use in the latexml code, constructors and such. This must be a one to one mapping and there are no default namespaces. Document: this namespace mapping associates prefixes to namespace URIs as used in the generated document, and will be the set of prefixes used in the generated output. This mapping may also use a prefix of “#default” which is for the unprefixed form of elements (not used for attributes!)

pub fn register_namespace_sym( &mut self, codeprefix: SymbolU32, namespace_opt: Option<SymbolU32> )

pub fn register_document_namespace( &mut self, docprefix: &str, namespace_opt: Option<&str> )

pub fn set_relaxng_schema(&mut self, schema: &str)

pub fn load_compiled_schema(&mut self, path: &str)

TODO: This is another component that would fit perfectly as a compiler plugin. For now, simply reimplementing the runtime loading of LaTeXML.model as-is from Model.pm

pub fn add_tag_content(&mut self, tag: &str, elements: Vec<&str, Global>)

pub fn add_tag_attribute(&mut self, tag: &str, attributes: Vec<&str, Global>)

pub fn set_schema_class( &mut self, classname: &str, content: HashSet<SymbolU32, BuildHasherDefault<FxHasher>> )

pub fn describe_model(&self)