Module rtx_package::package::pathname

helper methods for file system paths

Structs

PathnameFindOptions

Functions

absolute

candidate_pathnames

It’s presumably cheep to concatinate all the pathnames, relative to the cost of testing for files, and this simplifies overall.

canonical

This likely needs portability work!!! (particularly regarding urls, separators, …) AND, care about symbolic links and collapsing ../ !!!

concat

cwd

directory

extension

file_name

file_stem

find

is_absolute

is_literaldata

is_nasty

is_reloadable

check whether a pathname is reloadable as a TeX definition

is_url

kpsewhich

protocol

Note that this returns ONLY recognized protocols!

split

Split the pathname into components (dir,name,type). If pathname is absolute, dir starts with volume or ‘/’

url_split

Simple logic for splitting a URL into protocol://base/path