Type Definition rtx_package::package::state::Table

pub type Table = HashMap<String, VecDeque<Stored, Global>, RandomState>;

The State efficiently maintain the bindings in a TeX-like fashion. bindings associate data with keys (eg definitions with macro names) and respect TeX grouping; that is, an assignment is only in effect until the current group (opened by \bgroup) is closed (by \egroup).

The objective is to make the following, most-common, operations FAST: begin & end a group (ie. push/pop a stack frame) lookup & assignment of values With the more obvious approach, a “stack of frames”, either lookup would involve checking a sequence of frames until the current value is found; or, starting a new frame would involve copying bindings for all values I never quite studied how Knuth does it; The following structures allow these to be constant operations (usually), except for endgroup (which is linear in # of changed values in that frame).

There are 2 main structures used here. For each of several Tables (being “value”, “meaning”, “catcode” or other space of names), each table maintains the bound values, and “undo” defines the stack frames: self.table[key] = [current_value, previous_value, …] self.undo[frame].table[key] = (None | n) such that the “current value” associated with key is the 0th element of the table array; the previous_values (if any) are values that had been assigned within previous groups. The undo list indicates how many values have been assigned for key in the frame’th frame (usually the last is the one of interest). [Would be simpler to store boolean in undo, but see deactivateScope] [An UndoFrame contains fields for each State table, and a lock attribute]

So, in handwaving form, the algorithms are as follows: push-frame == bgroup == begingroup: push an empty hash {} onto the undo stack; pop-frame == egroup == endgroup: for the n associated with every key in the topmost hash in the undo stack pop n values from the table then remove the hash from the undo stack. Lookup value: we simply fetch the last element from the table Assign a value: local scope (the normal way): we push a new value into the table described above, and also increment the associated value in the undo stack global scope: remove any locally scoped values, and undo entries for the key then set the only remaining value to the given one. named-scope scope: push an entry [table,key,value] globally to the stash table’s value. And assign locally, if the scope is active (has non-zero value in stash_active table),

There are tables for catcode: keys are char; Also, math:char =1 when char is active in math. mathcode, sfcode, lccode, uccode, delcode : are similar to catcode but store additional kinds codes per char (see TeX) value: keys are anything (typically a string, though) and value is the value associated with it meaning: The definition assocated with key, usually a control-sequence. stash & stash_active: support named scopes (see also activateScope & deactivateScope)