Struct rtx_package::package::ConstructorOptions

pub struct ConstructorOptions {

    pub nargs: Option<usize>,
    pub bounded: bool,
    pub mode: Option<String>,
    pub sizer: Option<Rc<dyn Fn(&Whatsit) -> Result<(Dimension, Dimension, Dimension), Error>>>,
    pub before_digest: Vec<Rc<dyn Fn() -> Result<Vec<Digested>, Error>>>,
    pub after_digest: Vec<Rc<dyn Fn(&mut Whatsit) -> Result<Vec<Digested>, Error>>>,
    pub before_construct: Vec<Rc<dyn Fn(&mut Document, &Whatsit) -> Result<(), Error>>>,
    pub after_construct: Vec<Rc<dyn Fn(&mut Document, &Whatsit) -> Result<(), Error>>>,
    pub require_math: bool,
    pub forbid_math: bool,
    pub properties: Rc<dyn Fn(&Vec<Option<Digested>>) -> Result<HashMap<String, Stored, BuildHasherDefault<FxHasher>>, Error>>,
    pub capture_body: bool,
    pub font: Option<FontDirective>,
    pub after_digest_begin: Vec<Rc<dyn Fn(&mut Whatsit) -> Result<Vec<Digested>, Error>>>,
    pub before_digest_end: Vec<Rc<dyn Fn() -> Result<Vec<Digested>, Error>>>,
    pub after_digest_body: Vec<Rc<dyn Fn(&mut Whatsit) -> Result<Vec<Digested>, Error>>>,
    pub reversion: Option<Reversion>,
    pub scope: Option<Scope>,
    pub robust: bool,
    pub locked: bool,
    pub alias: Option<String>,

}

configuration for creating a new Constructor

Fields
nargs: Option<usize>

number of arguments (if any)

bounded: bool

bouded mode (default: false)

mode: Option<String>

begin a named mode

sizer: Option<Rc<dyn Fn(&Whatsit) -> Result<(Dimension, Dimension, Dimension), Error>>>

a SizingClosure to estimate the size of the digested box

before_digest: Vec<Rc<dyn Fn() -> Result<Vec<Digested>, Error>>>

custom code to run immediately before the digestion phase

after_digest: Vec<Rc<dyn Fn(&mut Whatsit) -> Result<Vec<Digested>, Error>>>

custom code to run immediately after the digestion phase

before_construct: Vec<Rc<dyn Fn(&mut Document, &Whatsit) -> Result<(), Error>>>

custom code to run immediately before the construction phase

after_construct: Vec<Rc<dyn Fn(&mut Document, &Whatsit) -> Result<(), Error>>>

custom code to run immediately after the construction phase

require_math: bool

requires to be used in math mode

forbid_math: bool

forbids use in math mode

properties: Rc<dyn Fn(&Vec<Option<Digested>>) -> Result<HashMap<String, Stored, BuildHasherDefault<FxHasher>>, Error>>

custom directives for computing box properties

capture_body: bool

should it capture the body as #body (default: false)

font: Option<FontDirective>

specify a font to use, or instructions how to compute which font to use

after_digest_begin: Vec<Rc<dyn Fn(&mut Whatsit) -> Result<Vec<Digested>, Error>>>

custom code to run as digestion begins

before_digest_end: Vec<Rc<dyn Fn() -> Result<Vec<Digested>, Error>>>

custom code to run just before digestion ends

after_digest_body: Vec<Rc<dyn Fn(&mut Whatsit) -> Result<Vec<Digested>, Error>>>

custom code to run after #body has been digested

reversion: Option<Reversion>

provide tokens to revert to, or custom code for computing them

scope: Option<Scope>

Local/Global scope of installing this definition (default: Local)

robust: bool

is this a robust command sequence (default: false)

locked: bool

lock the definition for raw TeX overrides (default: false)

alias: Option<String>

alternative (command sequence) name, used for reversion