Struct rtx_package::package::Document

pub struct Document {
    pub document: Document,
    pub pending: Vec<Node, Global>,
    pub node: Node,
    pub node_boxes: HashMap<usize, Digested, BuildHasherDefault<FxHasher>>,
    pub node_fonts: HashMap<u64, Font, BuildHasherDefault<FxHasher>>,
    pub idstore: HashMap<String, Node, BuildHasherDefault<FxHasher>>,
    pub rewrite_labels: HashMap<String, String, BuildHasherDefault<FxHasher>>,
    /* private fields */
}

Fields
document: Documentpending: Vec<Node, Global>node: Nodenode_boxes: HashMap<usize, Digested, BuildHasherDefault<FxHasher>>node_fonts: HashMap<u64, Font, BuildHasherDefault<FxHasher>>idstore: HashMap<String, Node, BuildHasherDefault<FxHasher>>rewrite_labels: HashMap<String, String, BuildHasherDefault<FxHasher>>
Implementations

impl Document

pub fn new() -> Document

pub fn get_element(&self) -> Option<Node>

Get the element at (or containing) the current insertion point.

pub fn findnodes( &mut self, xpath: &str, node_opt: Option<&Node>, state: &mut State ) -> Vec<Node, Global>

Find the nodes according to the given xpath expression, the xpath is relative to $node (if given), otherwise to the document node.

pub fn get_xpath(&mut self, model: &Model) -> &mut XPath

Get an XPath context that knows about our namespace mappings.

pub fn findnode( &mut self, xpath: &str, node: Option<&Node>, state: &mut State ) -> Option<Node>

Like findnodes, but only returns the first matched node

pub fn findvalues( &mut self, xpath: &str, node_opt: Option<&Node>, state: &mut State ) -> Vec<String, Global>

Like findnodes, but expects an xpath that evaluates to a literal value (e.g. for attributes)

pub fn get_node_qname(&self, node: &Node, state: &State) -> SymbolU32

Get the node’s qualified name in standard form Ie. using the registered prefix for that namespace. NOTE: Reconsider how Capture & WildCard should be integrated!?! NOTE: Should Deprecate! (use model)

pub fn with_node_qname<R, FnR>( &self, node: &Node, state: &State, caller: FnR ) -> Rwhere FnR: FnOnce(&str) -> R,

pub fn get_node(&self) -> &Node

pub fn get_node_mut(&mut self) -> &mut Node

pub fn get_document(&self) -> &Document

pub fn get_document_mut(&mut self) -> &mut Document

pub fn finalize(&mut self, state: &mut State) -> Result<(), Error>

pub fn absorb( &mut self, object: &Digested, props_opt: Option<HashMap<String, Stored, BuildHasherDefault<FxHasher>>>, state: &mut State ) -> Result<(), Error>

Document construction at the Current Insertion Point.

absorb the given $box into the DOM (called from constructors). This will return a list of whatever nodes were created. Note that this may include nodes that are children of other nodes in the list or nodes that are no longer in the document. Also, note that when a text nodes is appended to, the complete text node is in the list, not just the portion that was added. [Note that recording the nodes being constructed isn’t all that costly, but filtering them for parent/child relations IS, particularly since it usually isn’t needed]

A box that is a TBox, or List, or Whatsit, is responsible for carrying out its own insertion, but it should ultimately call methods of Document that will record the nodes that were created. $box can also be a plain string (Digested::Postponed) which will be inserted according to whatever font, mode, etc, are in %props.

pub fn get_constructed_nodes(&self) -> &[Node]

pub fn absorb_string( &mut self, object: &str, props: &HashMap<String, Stored, BuildHasherDefault<FxHasher>>, state: &mut State ) -> Result<Option<Node>, Error>

This is a refactored else cases from the main absorb routine, to allow for better type hygiene

pub fn insert_element( &mut self, qname: &str, content: Vec<&Digested, Global>, attrib: Option<HashMap<String, String, BuildHasherDefault<FxHasher>>>, state: &mut State ) -> Result<Node, Error>

Shorthand for open,absorb,close, but returns the new node.

pub fn insert_pi( &mut self, op: &str, attributes_opt: Option<HashMap<String, String, BuildHasherDefault<FxHasher>>> ) -> Result<(), Error>

Insert a ProcessingInstruction of the form Does NOT move the current insertion point to the PI, but may move up past a text node.

pub fn open_element( &mut self, qname: &str, attributes: Option<HashMap<String, String, BuildHasherDefault<FxHasher>>>, font_opt: Option<&Font>, state: &mut State ) -> Result<Node, Error>

pub fn close_element( &mut self, qname: &str, state: &mut State ) -> Result<Option<Node>, Error>

Note: This closes the deepest open node of a given type. This can cause problems with auto-opened nodes, esp. ones for fontswitches! Since this is an “explicit request”, we’re currently skipping over those nodes, ie. we’re automatically closing them, even if they’re the same type as we’re asking to close!!! This is kinda risky! Maybe we should try to request closing of specific nodes.

pub fn is_openable(&self, qname: &str, state: &mut State) -> bool

pub fn is_closeable<T>(&self, tags: T, state: &mut State) -> Option<Node>where T: IntoVDQS,

Check whether it is possible to close each element in @tags, any intervening nodes must be autocloseable. returning the last Some(node) that would be closed if it is possible, otherwise None

pub fn maybe_close_element( &mut self, qname: &str, state: &mut State ) -> Result<Option<Node>, Error>

pub fn close_to_node( &mut self, node: &Node, _ifopen: bool, state: &mut State ) -> Result<(), Error>

Closes all nodes until $node becomes the current point.

pub fn close_node( &mut self, node: &Node, state: &mut State ) -> Result<(), Error>

Closes all nodes until $node is closed.

pub fn maybe_close_node( &mut self, node: &Node, state: &mut State ) -> Result<(), Error>

Only if needed/possible: closes all nodes until $node is closed

pub fn close_node_with_strictness( &mut self, strict: bool, node: &Node, state: &mut State ) -> Result<(), Error>

pub fn can_auto_close(&self, node: &Node, state: &State) -> bool

pub fn get_tag_action_list( &self, tag: SymbolU32, when: TagOptionName, state: &mut State ) -> Vec<Rc<dyn Fn(&mut Document, &mut Node, &mut State) -> Result<(), Error>>, Global>

get the actions that should be performed on afterOpen or afterClose

pub fn serialize_to_string(&self, state: &mut State) -> String

pub fn serialize_aux( &self, node: &Node, depth: usize, noindent: bool, heuristic: bool, state: &mut State ) -> String

We ought to try for something close to C14N (http://www.w3.org/TR/xml-c14n), but keep XML declaration, comments and don’t convert empty elements.

pub fn set_node(&mut self, node: &Node)

pub fn insert_math_token( &mut self, text: &str, attributes: HashMap<String, String, BuildHasherDefault<FxHasher>>, font_opt: Option<&Font>, state: &mut State ) -> Result<Node, Error>

pub fn insert_comment( &mut self, text: &str, state: &mut State ) -> Result<Node, Error>

Insert a new comment, or append to previous comment. Does NOT move the current insertion point to the Comment, but may move up past a text node.

pub fn open_text( &mut self, text: &str, font: &Font, state: &mut State ) -> Result<Option<Node>, Error>

pub fn can_contain(&self, node: &Node, child: &str, state: &mut State) -> bool

pub fn can_contain_qname( &self, tag: &str, child: &str, state: &mut State ) -> bool

pub fn can_contain_sym( &self, node: &Node, child: SymbolU32, state: &mut State ) -> bool

pub fn can_contain_qsym( &self, tag: SymbolU32, child: SymbolU32, state: &mut State ) -> bool

pub fn can_contain_indirect( &self, tag: SymbolU32, child: SymbolU32, state: &mut State ) -> Option<SymbolU32>

Can an element with (qualified name) $tag contain a $childtag element indirectly? That is, by openning some number of autoOpen’able tags? And if so, return the tag to open.

pub fn can_contain_node_somehow( &self, node: &Node, child: &str, state: &mut State ) -> bool

pub fn can_contain_somehow( &self, tag: &str, child: &str, state: &mut State ) -> bool

pub fn sym_can_contain_somehow( &self, tag: SymbolU32, child: SymbolU32, state: &mut State ) -> bool

pub fn can_node_have_attribute( &mut self, node: &Node, attrib: &str, state: &mut State ) -> bool

pub fn can_have_attribute( &self, tag: &str, attrib: &str, state: &mut State ) -> bool

pub fn sym_can_have_attribute( &self, tag: SymbolU32, attrib: SymbolU32, state: &mut State ) -> bool

pub fn close_text_internal(&mut self, state: &State) -> Result<Node, Error>

pub fn close_node_internal( &mut self, node: &Node, state: &mut State ) -> Result<(), Error>

Close node, and any current nodes below it. No checking! Use this when you’ve already verified that node can be closed. and, of course, node must be current or some ancestor of it!!!

pub fn merge_attributes( &mut self, from: Node, to: &mut Node, _force: Option<HashSet<&'static str, BuildHasherDefault<FxHasher>>> ) -> Result<(), Error>

pub fn open_text_internal( &mut self, text: &str, state: &mut State ) -> Result<Node, Error>

pub fn record_constructed_node(&mut self, node: &Node)

Note that a box has been absorbed creating node; This does book keeping so that we can return the sequence of nodes that were added by absorbing material.

pub fn filter_deletions(&self, nodes: Vec<Node, Global>) -> Vec<Node, Global>

pub fn filter_children(&self, nodes: Vec<Node, Global>) -> Vec<Node, Global>

Given a list of nodes such as from ->absorb, filter out all the nodes that are children of other nodes in the list.

pub fn get_insertion_context( &self, levels_opt: Option<usize>, state: &mut State ) -> Result<String, Error>

Return a string indicating the path to the current insertion point in the document. if $levels is defined, show only that many levels

pub fn find_insertion_point( &mut self, qname: &str, has_opened_opt: Option<SymbolU32>, state: &mut State ) -> Result<Node, Error>

Find the node where an element with qualified name $qname can be inserted. This will move up the tree (closing auto-closable elements), or down (inserting auto-openable elements), as needed.

pub fn node_set_attribute( &mut self, key: &str, value: &str, state: &mut State ) -> Result<(), Error>

pub fn node_get_attribute(&mut self, name: &str) -> Option<String>

pub fn set_attribute( &mut self, node: &mut Node, key: &str, value: &str ) -> Result<(), Error>

pub fn add_ss_values( &mut self, node: &mut Node, key: &str, values_str: &str ) -> Result<(), Error>

pub fn add_class(&mut self, node: &mut Node, class: &str) -> Result<(), Error>

pub fn unrecord_id(&mut self, id: &str)

pub fn record_node_ids( &mut self, node: &Node, state: &mut State ) -> Result<(), Error>

These are used to record or unrecord, in bulk, all the ids within a node (tree).

pub fn unrecord_node_ids(&mut self, node: &Node, state: &mut State)

pub fn modify_id(&mut self, id: String) -> String

Get a new, related, but unique id Sneaky option: try “ID_SUFFIX” as a suffix for id, first.

pub fn lookup_id(&self, id: &str) -> Option<&Node>

pub fn mark_xmnode_visibility(&mut self, state: &mut State) -> Result<(), Error>

pub fn set_node_box(&mut self, node: &Node, digested: Digested)

Record the Box that created this node.

pub fn get_node_box(&self, node: &Node) -> Option<Digested>

pub fn set_node_font( &mut self, node: &mut Node, font: &Font ) -> Result<(), Error>

Record the Font of a node

pub fn copy_node_font( &mut self, from: &Node, to: &mut Node ) -> Result<(), Error>

pub fn merge_node_font_rec( &mut self, node: &Node, font: &Font ) -> Result<(), Error>

Possibly a sign of a design flaw; Set the node’s font & all children that HAD the same font.

pub fn set_box_font( &mut self, node: &mut Node, state: &mut State ) -> Result<(), Error>

pub fn get_node_font(&self, node: &Node) -> &Font

pub fn has_node_font(&self, node: &Node) -> bool

pub fn get_node_language(&self, node: &Node) -> String

pub fn remove_node(&mut self, node: &mut Node)

pub fn open_element_at( &mut self, point: &mut Node, qname: &str, attributes: Option<HashMap<String, String, BuildHasherDefault<FxHasher>>>, font_opt: Option<Font>, state: &mut State ) -> Result<Node, Error>

This opens a new element at the specified point, rather than the current insertion point. This is useful during document rearrangement or augmentation that may be needed later in the process.

pub fn close_element_at( &mut self, node: &mut Node, state: &mut State ) -> Result<(), Error>

Whenever a node has been created using openElementAt, closeElementAt ought to be used to close it, when you’re finished inserting into $node. Basically, this just runs any afterClose operations.

pub fn after_open( &mut self, node: &mut Node, state: &mut State ) -> Result<(), Error>

pub fn after_close( &mut self, node: &mut Node, state: &mut State ) -> Result<(), Error>

pub fn append_clone( &mut self, node: &mut Node, new_children: Vec<Node, Global>, state: &mut State ) -> Result<(), Error>

pub fn wrap_nodes( &mut self, qname: &str, nodes: Vec<Node, Global>, state: &mut State ) -> Result<Option<Node>, Error>

pub fn unwrap_nodes(&mut self, node: Node) -> Result<Node, Error>

Unwrap the children of $node, by replacing $node by its children.

pub fn replace_node( &mut self, node: Node, with: Vec<Node, Global> ) -> Result<Node, Error>

pub fn rename_node( &mut self, _node: &mut Node, _newname: &str ) -> Result<Node, Error>

pub fn trim_node_whitespace(&mut self, node: &mut Node) -> Result<(), Error>

pub fn add_resource( &mut self, resource: Resource, state: &mut State ) -> Result<(), Error>

pub fn process_pending_resources( &mut self, state: &mut State ) -> Result<(), Error>

pub fn make_error( &mut self, error_class: &str, _content: &str, state: &mut State ) -> Result<(), Error>

pub fn float_to_element( &mut self, qname: &str, closeifpossible: bool, state: &mut State ) -> Result<Option<Node>, Error>

Find a node in the document that can contain an element qname

pub fn float_to_label(&mut self, state: &mut State) -> Option<Node>

pub fn set_box_to_absorb(&mut self, arg: Option<Digested>)

pub fn expire_box_to_absorb(&mut self)

pub fn load_labels_for_rewrite( &mut self, state: &mut State ) -> Result<(), Error>

pub fn generate_id( &mut self, node: &mut Node, prefix: &str, state: &mut State ) -> Result<(), Error>

This function computes an xml:id for a node, if it hasn’t already got one. It is suitable for use in Tag afterOpen as Tag('ltx:para',afterOpen=>sub { GenerateID(@_,'p'); }); It generates an id of the form . The parent node (the one with ID=) also maintains a counter stored in an attribute _ID_counter_<prefix> recording the last used for amongst its descendents.

pub fn replace_tree( &mut self, new: Node, old: &mut Node, state: &mut State ) -> Result<Option<Node>, Error>

pub fn append_tree( &mut self, node: &mut Node, data: Vec<Node, Global>, state: &mut State ) -> Result<(), Error>