Struct rtx_package::package::Float

pub struct Float(pub f64);

Tuple Fields
0: f64
Implementations

impl Float

pub fn to_tight_string(&self) -> String

Tight formatting of floats, where we emit them as integers when they do not have a decimal part used in e.g. the multido.sty binding and test