Struct rtx_package::package::Font

pub struct Font {

    pub family: Option<Cow<'static, str>>,
    pub series: Option<Cow<'static, str>>,
    pub shape: Option<Cow<'static, str>>,
    pub size: Option<f64>,
    pub color: Option<Cow<'static, str>>,
    pub bg: Option<Cow<'static, str>>,
    pub opacity: Option<Cow<'static, str>>,
    pub encoding: Option<Cow<'static, str>>,
    pub language: Option<Cow<'static, str>>,
    pub mathstyle: Option<Cow<'static, str>>,
    pub mathstylestep: Option<i32>,
    pub name: Option<Cow<'static, str>>,
    pub emph: Option<bool>,
    pub scripted: Option<bool>,
    pub forceseries: Option<bool>,
    pub forcefamily: Option<bool>,
    pub forceshape: Option<bool>,
    pub scale: Option<f64>,
    pub flags: Option<u8>,

}

This struct is a little interesting, as we want to pass overrides that partially modify (via a merge) the current font, in each definitional binding. To accommodate that with this struct, every single field needs to be an Option, in order to unambiguously tell the “intend” of override (Some) vs no intent (None).

Fields
family: Option<Cow<'static, str>>series: Option<Cow<'static, str>>shape: Option<Cow<'static, str>>size: Option<f64>color: Option<Cow<'static, str>>bg: Option<Cow<'static, str>>opacity: Option<Cow<'static, str>>encoding: Option<Cow<'static, str>>language: Option<Cow<'static, str>>mathstyle: Option<Cow<'static, str>>mathstylestep: Option<i32>name: Option<Cow<'static, str>>emph: Option<bool>scripted: Option<bool>forceseries: Option<bool>forcefamily: Option<bool>forceshape: Option<bool>scale: Option<f64>flags: Option<u8>
Implementations

impl Font

pub fn text_default() -> Font

pub fn math_default() -> Font

pub fn to_hashable(&self) -> u64

pub fn math_bearing(&self, _thisbox: &Digested, _prevbox: &Digested) -> f64

pub fn is_sticky(&self) -> bool

pub fn get_family(&self) -> Option<&Cow<'_, str>>

pub fn get_series(&self) -> Option<&Cow<'_, str>>

pub fn get_shape(&self) -> Option<&Cow<'_, str>>

pub fn get_size(&self) -> Option<f64>

pub fn get_color(&self) -> Option<&Cow<'_, str>>

pub fn get_background(&self) -> Option<&Cow<'_, str>>

pub fn get_opacity(&self) -> Option<&Cow<'_, str>>

pub fn get_encoding(&self) -> Option<&Cow<'_, str>>

pub fn get_language(&self) -> Option<&Cow<'_, str>>

pub fn get_mathstyle(&self) -> Option<&Cow<'_, str>>

pub fn get_flags(&self) -> Option<u8>

pub fn merge(&self, other: Font) -> Font

pub fn specialize(&self, text: &str) -> Font

Instanciate the font for a particular class of symbols. NOTE: This works in `normal’ latex, but probably needs some tunability. Depending on the fonts being used, the allowable combinations may be different. Getting the font right is important, since the author probably thinks of the identity of the symbols according to what they SEE in the printed document. Even though the markup might seem to indicate something else…

Use Unicode properties to determine font merging.

pub fn distance(&self, other: &Font) -> i8

pub fn relative_to(
&self,
other: &Font
) -> HashMap<String, (String, Font), BuildHasherDefault<FxHasher>>

This method compares 2 fonts, returning the differences between them. Noting that the font-related attributes in the schema distill the font properties into fewer attributes (font,fontsize,color,background,opacity), the return value encodes both the attribute changes that would be needed to effect the font change, along with the font properties that differed Namely, the result is a hash keyed on the attribute name and whose value is a FontDiff value => “new_attribute_value” properties => { %fontproperties } or (String, Font)

pub fn purestyle_changes(&self, other: &Font) -> Font

pub fn get_metric(&self, c_opt: Option<char>) -> &MetricData

pub fn get_em_width(&self) -> i64

pub fn get_ex_height(&self) -> i64

pub fn get_mu_width(&self) -> i64

pub fn compute_string_size(
&self,
text: &str,
_options: HashMap<String, Stored, BuildHasherDefault<FxHasher>>,
_state: &State
) -> (Dimension, Dimension, Dimension)

pub fn compute_boxes_size(
&self,
boxes: &[Digested],
options: HashMap<String, Stored, BuildHasherDefault<FxHasher>>,
state: &mut State
) -> Result<(Dimension, Dimension, Dimension), Error>