Struct rtx_package::package::Mouth

pub struct Mouth { /* private fields */ }

Implementations

impl Mouth

pub fn create(
source: &str,
options: MouthOptions,
state: &mut State
) -> Result<Mouth, Error>

pub fn new(
text: &str,
options: Option<MouthOptions>,
state: &mut State
) -> Result<Mouth, Error>

pub fn get_source(&self) -> &str

pub fn open(&mut self, content: &str, state: &mut State) -> Result<(), Error>

pub fn finish(&mut self, state: &mut State)

pub fn has_more_input(&mut self) -> bool

pub fn read_token(&mut self, state: &mut State) -> Option<Token>

Read the next token, or undef if exhausted. Note that this also returns COMMENT tokens containing source comments, and also locator comments (file, line# info). LaTeXML::Core::Gullet intercepts them and passes them on at appropriate times.

pub fn read_tokens(&mut self, state: &mut State) -> Tokens

Read all tokens until a token equal to $until (if given), or until exhausted. Returns an empty Tokens list, if there is no input

pub fn read_raw_line(&mut self, noread: bool, state: &State) -> Option<String>

pub fn is_eol(&mut self, state: &State) -> bool

pub fn at_eof(&self) -> bool