Struct rtx_package::package::Parameter

pub struct Parameter {

    pub novalue: bool,
    pub semiverbatim: Option<Vec<char, Global>>,
    pub optional: bool,
    pub pack_parameters: bool,
    pub name: Cow<'static, str>,
    pub spec: Cow<'static, str>,
    pub extra: Vec<Tokens, Global>,
    pub inner: Option<Parameters>,
    pub reader: Rc<dyn Fn(&mut Gullet, Option<&Parameters>, &[Tokens], &mut State) -> Result<ArgWrap, Error> + 'static>,
    pub reader_predigest: Option<Rc<dyn Fn(&mut Stomach, ArgWrap, &mut State) -> Result<Option<Digested>, Error> + 'static>>,
    pub reversion: Option<Rc<dyn Fn(Vec<Token, Global>, Option<&Parameters>, &[Tokens], &State) -> Result<Tokens, Error> + 'static>>,
    pub before_digest: Vec<Rc<dyn Fn(&mut Stomach, &mut State) -> Result<Vec<Digested, Global>, Error> + 'static>, Global>,
    pub after_digest: Vec<Rc<dyn Fn(&mut Stomach, &mut Whatsit, &mut State) -> Result<Vec<Digested, Global>, Error> + 'static>, Global>,

}

Fields
novalue: boolsemiverbatim: Option<Vec<char, Global>>optional: boolpack_parameters: boolname: Cow<'static, str>spec: Cow<'static, str>extra: Vec<Tokens, Global>inner: Option<Parameters>reader: Rc<dyn Fn(&mut Gullet, Option<&Parameters>, &[Tokens], &mut State) -> Result<ArgWrap, Error> + 'static>reader_predigest: Option<Rc<dyn Fn(&mut Stomach, ArgWrap, &mut State) -> Result<Option<Digested>, Error> + 'static>>reversion: Option<Rc<dyn Fn(Vec<Token, Global>, Option<&Parameters>, &[Tokens], &State) -> Result<Tokens, Error> + 'static>>before_digest: Vec<Rc<dyn Fn(&mut Stomach, &mut State) -> Result<Vec<Digested, Global>, Error> + 'static>, Global>after_digest: Vec<Rc<dyn Fn(&mut Stomach, &mut Whatsit, &mut State) -> Result<Vec<Digested, Global>, Error> + 'static>, Global>
Implementations

impl Parameter

pub fn new( name: Cow<'static, str>, spec: Cow<'static, str>, state: &mut State ) -> Result<Parameter, Error>

pub fn init(self, state: &mut State) -> Result<Parameter, Error>

pub fn check_reader_function( name: &str, state: &State ) -> Option<Rc<dyn Fn(&mut Gullet, Option<&Parameters>, &[Tokens], &mut State) -> Result<ArgWrap, Error> + 'static>>

Obtain the reader of a given parameter name, if available

pub fn setup_catcodes(&self, state: &mut State)

pub fn revert_catcodes(&self, state: &mut State) -> Result<(), Error>

pub fn read( &self, gullet: &mut Gullet, fordefn: Option<&dyn Definition>, state: &mut State ) -> Result<ArgWrap, Error>

pub fn digest( &self, stomach: &mut Stomach, value_arg: ArgWrap, _fordefn: Option<&Constructor>, state: &mut State ) -> Result<Option<Digested>, Error>

pub fn revert( &self, value_opt: Option<Tokens>, state: &State ) -> Result<Option<Tokens>, Error>

pub fn reparse( &self, _value: Tokens, _gullet: &mut Gullet, _state: &State ) -> Result<Tokens, Error>

This is needed by structured parameter types like KeyVals where the argument may already have been tokenized before the KeyVals (and the parameter types for the keys) had a chance to properly parse.