Struct rtx_package::package::State

pub struct State {

    pub model: Model,
    pub tag_properties: HashMap<String, TagOptions, BuildHasherDefault<FxHasher>>,
    pub indirect_model: Option<HashMap<String, HashMap<String, String, BuildHasherDefault<FxHasher>>, BuildHasherDefault<FxHasher>>>,
    pub pending_resources: Vec<Resource, Global>,
    pub verbosity: i32,
    pub align_group_count: i32,
    pub status_code: usize,
    pub unlocked: bool,
    pub input_encoding: Option<String>,
    pub search_paths: VecDeque<String, Global>,
    pub graphics_paths: VecDeque<String, Global>,
    pub nomathparse: bool,
    pub reading_alignment: bool,
    pub bindings_dispatch: Option<Arc<dyn Fn(&str, &mut Stomach, &mut State) -> Option<Result<(), Error>> + 'static>>,
    pub extra_bindings_dispatch: Option<Arc<dyn Fn(&str, &mut Stomach, &mut State) -> Option<Result<(), Error>> + 'static>>,
    pub stomach: Arc<RwLock<Stomach>>,
    /* private fields */

}

The State efficiently maintain the bindings in a TeX-like fashion. bindings associate data with keys (eg definitions with macro names) and respect TeX grouping; that is, an assignment is only in effect until the current group (opened by \bgroup) is closed (by \egroup).

Fields
model: Model

the schema-derived model used for the current document

tag_properties: HashMap<String, TagOptions, BuildHasherDefault<FxHasher>>indirect_model: Option<HashMap<String, HashMap<String, String, BuildHasherDefault<FxHasher>>, BuildHasherDefault<FxHasher>>>

an optional indirect model for long-distance relationships

pending_resources: Vec<Resource, Global>

Document-related resources declared during core conversion, pending until XML is finalized

verbosity: i32align_group_count: i32status_code: usizeunlocked: boolinput_encoding: Option<String>search_paths: VecDeque<String, Global>

current paths to search for TeX inputs

graphics_paths: VecDeque<String, Global>

current paths to search for graphics

nomathparse: bool

flag to disable math parsing

reading_alignment: bool

TODO: marker for alignment

bindings_dispatch: Option<Arc<dyn Fn(&str, &mut Stomach, &mut State) -> Option<Result<(), Error>> + 'static>>

A dispatcher routing to the compiled code of the main/official rtx bindings

extra_bindings_dispatch: Option<Arc<dyn Fn(&str, &mut Stomach, &mut State) -> Option<Result<(), Error>> + 'static>>

Auxiliary convenience – extra dispatch

stomach: Arc<RwLock<Stomach>>

Circular dependency and global $STATE in Perl requires a bad style use of interior mutability…

Implementations

impl State

pub fn new(options: StateOptions) -> State

pub fn table(
&self,
name: TableName
) -> &HashMap<String, VecDeque<Stored, Global>, BuildHasherDefault<FxHasher>>

borrow/get the named table

pub fn table_mut(
&mut self,
name: TableName
) -> &mut HashMap<String, VecDeque<Stored, Global>, BuildHasherDefault<FxHasher>>

mutably borrow/get the named table

pub fn lookup_value(&self, key: &str) -> Option<&Stored>

fetches a Stored value at the given key, from the Value table

pub fn lookup_value_mut<'lv>(
&'lv mut self,
key: &'lv str
) -> Option<&'lv mut Stored>

mutably borrows a Stored value at the given key, from the Value table

pub fn remove_value<'lv>(&'lv mut self, key: &'lv str) -> Option<Stored>

inline lookup_value after which globally assign an empty Tokens() to undo

pub fn checkout_value(&mut self, key: &str) -> Option<Stored>

Replaces the value in question with Stored::None (see checkin_value for returning it)

pub fn checkin_value(&mut self, key: &str, value: Stored)

Returns a value into its Stored::None placeholder (see checkout_value for taking it)

pub fn assign_value<T, S, 'av>(&'av mut self, key: &'av str, value: T, scope: S)where
T: Into<Stored>,
S: Into<Option<Scope>>,

assigns a Stored value at the given key and scope

pub fn push_value<T>(&mut self, key: &str, value: T)where
T: Into<Stored>,

manage a (global) list of values

pub fn pop_value(&mut self, key: &str) -> Option<Stored>

pops the last value in a named Stored::VecDequeStored queue, if any

pub fn has_value(&self, key: &str) -> bool

Check if the Value table contains a given key

pub fn push_tokens(&mut self, key: &str, value: Tokens)

Pushes Tokens into a Stored::Tokens value when defined, or assigns when new.

pub fn lookup_bool(&self, key: &str) -> bool

A bit of Perl “existence as truth” semantics mixed in with proper boolean lookup

pub fn lookup_string(&self, key: &str) -> String

like lookup_value, but casts the entry into a String (empty if None)

pub fn lookup_int(&self, key: &str) -> i64

like lookup_value but only recognizes Int, Bool and Number variants of Stored (default: 0)

pub fn lookup_vec_string<'lvec>(
&'lvec self,
key: &'lvec str
) -> Option<&'lvec Vec<String, Global>>

pub fn lookup_vecdeque<'lvdq>(
&'lvdq self,
key: &'lvdq str
) -> Option<&'lvdq VecDeque<Stored, Global>>

pub fn remove_vecdeque<'lvdq>(
&'lvdq mut self,
key: &'lvdq str
) -> Option<VecDeque<Stored, Global>>

pub fn lookup_font(&self) -> Option<Arc<Font>>

convenience method to lookup the current value at the “font” key

pub fn lookup_mathfont(&self) -> Option<Arc<Font>>

convenience method to lookup the current value at the “mathfont” key

pub fn assign_font(&mut self, font: Arc<Font>, scope: Option<Scope>)

a convenience method to globally asign a Font to the “font” key

pub fn lookup_number(&self, key: &str) -> Option<Number>

a variant of lookup_value that casts the value into Number

pub fn lookup_dimension(&self, key: &str) -> Option<Dimension>

a variant of lookup_value that casts the value into Dimension

pub fn lookup_glue(&self, key: &str) -> Option<Glue>

a variant of lookup_value that only recognizes a Stored::Glue

pub fn lookup_muglue(&self, key: &str) -> Option<MuGlue>

a variant of lookup_value that only recognizes a Stored::Glue

pub fn lookup_tokens(&self, key: &str) -> Option<Tokens>

a variant of lookup_value that casts the response into Tokens

pub fn lookup_token(&self, key: &str) -> Option<&Token>

a variant of lookup_value that only recognizes a Stored::Token

pub fn lookup_register(
&mut self,
cs: &str,
parameters: Vec<ArgWrap, Global>
) -> Option<RegisterValue>

pub fn lookup_expandable(
&self,
token: &Token,
toplevel: bool
) -> Option<Arc<dyn Definition + 'static>>

pub fn is_dont_expandable(&self, token: &Token) -> bool

Whether token must be wrapped as dont_expand

pub fn lookup_conditional(&self, token: &Token) -> Option<ConditionalType>

pub fn unshift_value<T>(&mut self, key: &str, values: Vec<T, Global>)where
T: Into<Stored>,

pub fn shift_value(&mut self, key: &str) -> Option<Stored>

pub fn lookup_mapping(&self, map: &str, key: &str) -> Option<&Stored>

manage a (global) hash of values

pub fn assign_mapping<T>(&mut self, map: &str, key: &str, value: Option<T>)where
T: Into<Stored>,

pub fn lookup_mapping_keys(&self, map: &str) -> Vec<&str, Global>

pub fn lookup_stacked_values(&self, key: &str) -> Vec<&Stored, Global>

pub fn is_value_bound(&self, key: &str, frame_opt: Option<usize>) -> bool

Was name bound? If frame is given, check only whether it is bound in that frame (0 is the topmost).

pub fn lookup_catcode(&self, c: char) -> Option<Catcode>

Lookup & assign a character’s Catcode

pub fn assign_catcode(&mut self, key: char, value: Catcode, scope: Option<Scope>)

assigns a Catcode for a given character

pub fn lookup_mathcode(&self, key: &str) -> Option<u16>

like lookup_catcode but targets Mathcode and its table

pub fn assign_mathcode<T, C, S>(&mut self, key: C, value: T, scope: S)where
T: Into<u16>,
C: Into<char>,
S: Into<Option<Scope>>,

like assign_catcode but targets Mathcode and its table

pub fn lookup_sfcode(&self, key: char) -> Option<u16>

like lookup_catcode but targets Sfcode and its table

pub fn assign_sfcode<T, C, S>(&mut self, key: C, value: T, scope: S)where
T: Into<u16>,
C: Into<char>,
S: Into<Option<Scope>>,

like assign_catcode but targets Sfcode and its table

pub fn lookup_lccode(&self, key: char) -> Option<u16>

like lookup_catcode but targets Lccode and its table

pub fn assign_lccode<T, C, S>(&mut self, key: C, value: T, scope: S)where
T: Into<u16>,
C: Into<char>,
S: Into<Option<Scope>>,

like assign_catcode but targets Lccode and its table

pub fn lookup_uccode(&self, key: char) -> Option<u16>

like lookup_catcode but targets Uccode and its table

pub fn assign_uccode<T, C, S>(&mut self, key: C, value: T, scope: S)where
T: Into<u16>,
C: Into<char>,
S: Into<Option<Scope>>,

like assign_catcode but targets Uccode and its table

pub fn lookup_delcode(&self, key: char) -> Option<u16>

like lookup_catcode but targets Delcode and its table

pub fn assign_delcode<T, C, S>(&mut self, key: C, value: T, scope: S)where
T: Into<u16>,
C: Into<char>,
S: Into<Option<Scope>>,

like assign_catcode but targets Delcode and its table

pub fn lookup_meaning(&self, token: &Token) -> Option<Cow<'_, Stored>>

Get the `Meaning’ of a token. For active control sequence’s this may give the definition object (if defined) or another token (if \let) or undef Any other token is returned as is.

pub fn assign_meaning<T>(
&mut self,
token: &Token,
meaning: T,
scope: Option<Scope>
)where
T: Into<Stored>,

$meaning should be a definition (for defining active control sequences) or another token, for \let

pub fn lookup_definition<'def>(
&'def self,
key: &'def Token
) -> Option<Arc<dyn Definition + 'static>>

used for expansion & various queries Since we’re not doing digestion here, we don’t need to handle mathactive, nor cs let to executable tokens This returns a definition object, or undef

pub fn lookup_definition_stored<'def>(
&'def self,
key: &'def Token
) -> Option<Stored>

Returns a definition as Stored so that one can call .read_arguments, which can’t be specialized during compile-time over a trait object Instead we’ll dispatch via Stored at runtime, to allow generic calls

pub fn lookup_register_definition(&self, key: &Token) -> Option<Arc<Register>>

A specialized version of lookup_definition for registers, since we can’t adequately perform multi-dispatch when we have a “Self: Sized” for the Definition trait object.

pub fn lookup_digestable_definition<'def>(
&'def mut self,
token: &'def Token
) -> Option<Stored>

Recognizes mathactive tokens in math mode and also looks for cs that have been let to other `executable’ tokens. Returns a definition object, or a “self inserting” token. Used for digestion.

pub fn install_definition<T>(&mut self, definition: T, scope: Option<Scope>)where
T: Into<Stored>,

And a shorthand for installing definitions

pub fn push_frame(&mut self)

Starts a new level of grouping. Note that this is lower level than C<\bgroup>;

pub fn pop_frame(&mut self) -> Result<(), Error>

Ends the current level of grouping. Note that this is lower level than \egroup;

pub fn get_frame_depth(&self) -> usize

Determine depth of group nesting created by {,},\bgroup,\egroup,\begingroup,\endgroup by counting all frames which are not Daemon frames (and thus don’t possess FRAME_LOCK). This may give incorrect results for some special environments (e.g. minipage)

pub fn begin_semiverbatim(&mut self, extraspecials: Option<&[char]>)

begins a semiverbatim frame, neutralizing the usual + requested characters

pub fn end_semiverbatim(&mut self) -> Result<(), Error>

end by just calling pop_frame

pub fn set_prefix(&mut self, prefix: &str)

Set one of the definition prefixes global, etc (only global matters!)

pub fn get_prefix(&self, prefix: &str) -> bool

gets the current value of a named prefix

pub fn clear_prefixes(&mut self)

clears the global prefixes

pub fn activate_scope(&mut self, scope: &str)

pub fn deactivate_scope(&mut self, scope: &str)

Removes any definitions that were associated with the named scope. Normally not needed, since a scopes definitions are locally bound anyway.

pub fn get_known_scopes(&self) -> Vec<&String, Global>

return all known named scopes

pub fn get_active_scopes(&self) -> Vec<&String, Global>

return the currently activated named scopes

pub fn convert_unit(&self, unit_arg: &str) -> f64

convert a unit name into a f64 scaling factor over sp

pub fn note_status(&self, _category: &str, _what: &str)

TODO

pub fn compute_indirect_model(
&mut self
) -> HashMap<String, HashMap<String, String, BuildHasherDefault<FxHasher>>, BuildHasherDefault<FxHasher>>

The indirect model includes all elements allowed as direct children, and all descendents of a node that can be inserted after auto_open-ing intermediate elements. This model therefor includes information from the Schema, as well as autoOpen information that may be introduced in binding files. [Thus it should NOT be modifying the Model object, which may cover several documents in `] $imodel{$tag}{$child} => $open means if in $tag, to open $child, we must first open $open

pub fn initialize_stomach(&mut self)

Initialize various stomach parameters, preload, etc.

pub fn let_i(
&mut self,
token1: &Token,
token2: Token,
scope: Option<Scope>,
gullet: &mut Gullet
)

Let macro setter

pub fn x_equals(&self, token1: &Token, token2: &Token) -> bool

XEquals check for two token arguments

pub fn generate_error_stub(
&mut self,
caller: &mut Gullet,
token: &Token
) -> Result<Token, Error>

Generate a stub definition for an undefined control-sequence, along with appropriate error messge.

pub fn generate_ligature_id(&mut self) -> usize

simple id generator for a ligature

pub fn after_assignment(&mut self, gullet: &mut Gullet)

run the accumulated directives from \afterassignment

pub fn set_ifframe(&mut self, if_frame: Option<Arc<RwLock<IfFrame>>>)

sets a (originally Perl-local) IfFrame that needs to be manually expired.

pub fn get_ifframe(&self) -> Option<Arc<RwLock<IfFrame>>>

retrieves the most recent (originally Perl-local) IfFrame

pub fn expire_ifframe(&mut self)

expires the most recent (originally Perl-local) IfFrame

pub fn set_smuggle_the(&mut self, smuggle_the: bool)

set special (localized) flag for “\the smuggling mode”; useful for expanded definitions

pub fn get_smuggle_the(&self) -> bool

get special (localized) flag for “\the smuggling mode”; useful for expanded definitions

pub fn expire_smuggle_the(&mut self)

expire special (localized) flag for “\the smuggling mode”; useful for expanded definitions

pub fn set_current_token(&mut self, token: Arc<Token>)

sets the (localized) current token. see Stomach::invoke_token

pub fn expire_current_token(&mut self)

expires the most recent (localized) current token.

pub fn get_current_token(&self) -> Option<Arc<Token>>

gets the (localized) current token