Module rtx_package::package::tex_functions

Structs

KVSpec

Support for Key / Value arguments.

Functions

aligning_environment

and_split

classify_box

cleanup_math

decode_math_char

dimension_to_spaces

a candidate for use by \hskip, \hspace, etc… ?

do_def

insert_block

This attempts to be a generalize vbox construction; It tries to figure out whether an ltx:inline-block or ltx:para is needed, and attempts to figure out whether sequences of the inserted content need to be explicitly wrapped in some kind of block element (presumably ltx:p). It returns the inserted inner blocks, whether or not they got wrapped by that ltx:inline-block; which it DOESN’T TELL YOU ABOUT!

keyvals_aux

lowercase_token

make_generic_message

neutralize_font

parse_def_parameters

predigest_box_contents

Reading a Box’s content is crucially dependent on invoking the “{” token and obtaining a digested result Hence it is always needed to pair read_box_contents with its stomach-level counterpart, predigest_box_contents

read_box_contents

reenter_text_mode

revert_spec

set_align_or_class

split_tokens

today

uppercase_token

writable_tokens

Converts $tokens to a string in the fashion of \message and others: doubles #, converts to string; optionally adds spaces after control sequences in the spirit of the B Book, “show_token_list” routine, in 292.