Function rtx_package::package::tex_functions::predigest_box_contents

pub fn predigest_box_contents(
    stomach: &mut Stomach,
    _tokens: ArgWrap,
    state: &mut State
) -> Result<Option<Digested>>

Reading a Box’s content is crucially dependent on invoking the “{” token and obtaining a digested result Hence it is always needed to pair read_box_contents with its stomach-level counterpart, predigest_box_contents