Struct rtx_package::package::tokens::Tokens

pub struct Tokens(_);

Tokens are a thin wrapper over a vector of Token objects usually read from a Mouth. They are usually treated as an immutable interface, an have to be consumed via .unlist() for access to the underlying data.

Implementations

impl Tokens

pub fn new(tokens: Vec<Token, Global>) -> Tokens

pub fn unlist(self) -> Vec<Token, Global>

Return a list of the tokens making up this Tokens

pub fn unlist_ref(&self) -> &Vec<Token, Global>

Return a reference to the tokens making up this Tokens

pub fn unlist_mut(&mut self) -> &mut Vec<Token, Global>

Return a mutable reference to the tokens making up this Tokens

pub fn is_empty(&self) -> bool

Are there any tokens at all contained in this Tokens object

pub fn len(&self) -> usize

Number of contained Token entries

pub fn revert(self) -> Vec<Token, Global>

Return a string containing the TeX form of the Tokens

pub fn to_number(&self) -> Number

to_number casts back to a parsed Number (usually via gullet.read_number) which had to be re-converted to a Tokens for reentering the expansion flow

pub fn to_dimension(&self) -> Dimension

to_dimension casts back to a parsed Dimension (usually via gullet.read_dimension) which had to be re-converted to a Tokens for reentering the expansion flow

pub fn to_glue(&self) -> Glue

to_glue casts back to a parsed Glue (usually via gullet.read_glue) which had to be re-converted to a Tokens for reentering the expansion flow

pub fn to_mu_glue(&self) -> MuGlue

to_mu_glue casts back to a parsed MuGlue (usually via gullet.read_mu_glue) which had to be re-converted to a Tokens for reentering the expansion flow

pub fn to_mu_dimension(&self) -> MuDimension

to_mu_dimension casts back to a parsed MuGlue (usually via gullet.read_mu_glue) which had to be re-converted to a Tokens for reentering the expansion flow

pub fn to_float(&self) -> Float

to_float casts back to a parsed Float (usually via gullet.read_float) which had to be re-converted to a Tokens for reentering the expansion flow

pub fn to_keyvals(&self, state: &State) -> KeyVals

to_keyvals casts back to a parsed KeyVals (usually via a KeyVals parameter type) which had to be re-converted to a Tokens for reentering the expansion flow

pub fn equals(&self, other: Tokens) -> bool

Methods for overloaded ops.

pub fn unwrap_or_default(self) -> Tokens

pub fn unwrap(&self) -> &Tokens

pub fn stringify(&self) -> String

pub fn be_digested(
self,
stomach: &mut Stomach,
state: &mut State
) -> Result<Digested, Error>

pub fn neutralize(self, extraspecials: &[char], state: &State) -> Tokens

pub fn is_balanced(&self) -> bool

pub fn substitute_parameters(&self, args: &[Option<Cow<'_, Tokens>>]) -> Tokens

pub fn without_dont_expand(self) -> Tokens

pub fn untex(&self) -> String

pub fn pack_parameters(self) -> Tokens