Trait rtx_package::package::BoxOps

pub trait BoxOps: Object {

    // Required methods
    fn be_absorbed(&self, document: &mut Document) -> Result<Vec<Node>, Error>;
    fn get_string(&self) -> Result<Cow<'_, str>, Error>;
    fn with_properties<R, FnR>(&self, caller: FnR) -> R
       where FnR: FnOnce(&HashMap<String, Stored, BuildHasherDefault<FxHasher>>) -> R;
    fn get_font(&self) -> Result<Option<Cow<'_, Font>>, Error>;
    fn compute_size(
        &self,
        options: HashMap<String, Stored, BuildHasherDefault<FxHasher>>
    ) -> Result<(Dimension, Dimension, Dimension), Error>;

    // Provided methods
    fn unlist(&self) -> Vec<Digested> { ... }
    fn unlist_ref(&self) -> Vec<Cow<'_, Digested>> { ... }
    fn be_absorbed_mut(
        &mut self,
        _document: &mut Document
    ) -> Result<Vec<Node>, Error> { ... }
    fn get_tokens(&self) -> Option<&Tokens> { ... }
    fn get_properties(
        &self
    ) -> &HashMap<String, Stored, BuildHasherDefault<FxHasher>> { ... }
    fn get_properties_mut(
        &mut self
    ) -> &mut HashMap<String, Stored, BuildHasherDefault<FxHasher>> { ... }
    fn set_property<T>(&mut self, key: &str, value: T)
       where T: Into<Stored> { ... }
    fn get_property(&self, key: &str) -> Option<Cow<'_, Stored>> { ... }
    fn get_property_string(&self, key: &str) -> String { ... }
    fn get_property_mut(&mut self, key: &str) -> Option<&mut Stored> { ... }
    fn has_property(&self, key: &str) -> bool { ... }
    fn get_property_bool(&self, key: &str) -> bool { ... }
    fn get_body(&self) -> Result<Option<Digested>, Error> { ... }
    fn set_font(&mut self, _font: Rc<Font>) { ... }
    fn set_width<T>(&mut self, width: T)
       where T: Into<Stored> { ... }
    fn get_width(
        &self,
        options: Option<HashMap<String, Stored, BuildHasherDefault<FxHasher>>>
    ) -> Result<Option<RegisterValue>, Error> { ... }
    fn set_height<T>(&mut self, width: T)
       where T: Into<Stored> { ... }
    fn get_height(&self) -> Option<RegisterValue> { ... }
    fn set_depth<T>(&mut self, width: T)
       where T: Into<Stored> { ... }
    fn get_depth(&self) -> Option<RegisterValue> { ... }
    fn get_size(
        &mut self,
        options: Option<HashMap<String, Stored, BuildHasherDefault<FxHasher>>>
    ) -> Result<(Dimension, Dimension, Dimension, Dimension, Dimension, Dimension), Error> { ... }
    fn compute_size_and_cache(
        &mut self,
        options: HashMap<String, Stored, BuildHasherDefault<FxHasher>>
    ) -> Result<(Dimension, Dimension, Dimension), Error> { ... }

}

Common operations for Box-like (digested) data

Required Methods

fn be_absorbed(&self, document: &mut Document) -> Result<Vec<Node>, Error>

absorb the current object into the Document XML - returning the corresponding nodes

fn get_string(&self) -> Result<Cow<'_, str>, Error>

build a string representation of the underlying digested data

fn with_properties<R, FnR>(&self, caller: FnR) -> R
where FnR: FnOnce(&HashMap<String, Stored, BuildHasherDefault<FxHasher>>) -> R,

execute a function using this object’s named properties

fn get_font(&self) -> Result<Option<Cow<'_, Font>>, Error>

gets the associated font, if any

fn compute_size( &self, options: HashMap<String, Stored, BuildHasherDefault<FxHasher>> ) -> Result<(Dimension, Dimension, Dimension), Error>

computes and returns the size of a box-like object

Provided Methods

fn unlist(&self) -> Vec<Digested>

If composite, unwrap into the contained digested objects (or return self)

fn unlist_ref(&self) -> Vec<Cow<'_, Digested>>

fn be_absorbed_mut( &mut self, _document: &mut Document ) -> Result<Vec<Node>, Error>

be_absorbed but with allowed side-effects on the carrier (for Alignment only)

fn get_tokens(&self) -> Option<&Tokens>

get the underlying tokens (preceding digestion)

fn get_properties( &self ) -> &HashMap<String, Stored, BuildHasherDefault<FxHasher>>

deprecated: get the map of named properties. This can not be usable as long as we have any data behind a RefCell wrapper. Use with_properties instead.

fn get_properties_mut( &mut self ) -> &mut HashMap<String, Stored, BuildHasherDefault<FxHasher>>

get a mutable reference to the map of named properties

fn set_property<T>(&mut self, key: &str, value: T)
where T: Into<Stored>,

set a named property (allows all Stored types for values)

fn get_property(&self, key: &str) -> Option<Cow<'_, Stored>>

get a single named property (with special “isSpace” check)

fn get_property_string(&self, key: &str) -> String

fn get_property_mut(&mut self, key: &str) -> Option<&mut Stored>

get a mutable reference to a single named property (does NOT have the “isSpace” check)

fn has_property(&self, key: &str) -> bool

checks if a property key has been set

fn get_property_bool(&self, key: &str) -> bool

obtains a boolean property value (false unless Stored::Bool)

fn get_body(&self) -> Result<Option<Digested>, Error>

obtains the “body” of a digested object which captured it

fn set_font(&mut self, _font: Rc<Font>)

sets an associated font

fn set_width<T>(&mut self, width: T)
where T: Into<Stored>,

sets a “width” property, for sizing

fn get_width( &self, options: Option<HashMap<String, Stored, BuildHasherDefault<FxHasher>>> ) -> Result<Option<RegisterValue>, Error>

gets the “width” property value, if any

fn set_height<T>(&mut self, width: T)
where T: Into<Stored>,

sets a “height” property value, for sizing

fn get_height(&self) -> Option<RegisterValue>

gets the “height” property value, if any

fn set_depth<T>(&mut self, width: T)
where T: Into<Stored>,

sets a “depth” property value, for sizing

fn get_depth(&self) -> Option<RegisterValue>

gets the “depth” property value, if any

fn get_size( &mut self, options: Option<HashMap<String, Stored, BuildHasherDefault<FxHasher>>> ) -> Result<(Dimension, Dimension, Dimension, Dimension, Dimension, Dimension), Error>

gets the box size as a triple of (width, height, depth) the generic implementation is immutable and will recompute the size on each call see Digested::get_size for a variant with interior mutability which caches the box size

fn compute_size_and_cache( &mut self, options: HashMap<String, Stored, BuildHasherDefault<FxHasher>> ) -> Result<(Dimension, Dimension, Dimension), Error>

computes and caches (via named properties) the size of a box-like object

Object Safety

This trait is not object safe.