Module rtx_package::prelude::binding::content

Structs

FindFileOptions

configuration for searching for a file in the local filesystem

InputDefinitionOptions

a configuration for loading LaTeX definition files (such as .sty, .cls, and their bindings)

InputOptions

configuration for input of a TeX source (content files mostly)

RequireOptions

Functions

build_invocation

Creates a single Tokens representing a TeX invocation of the lead token over a list of arguments.

convert_latex_args

Convert a LaTeX-style argument spec to our Package form. Ie. given $nargs and $optional, being the two optional arguments to something like \newcommand, convert it to the form we use

digest_if

digest_literal

digest_text

find_file

search for a file as prescribed by a FindFileOptions configuration

input

This is essentially the \input equivalent; we are most likely expecting to get actual content, (possibly with definitions included, as well) but might actually be getting pure definitions, (like a proper style file) in which case we may really want to load a binding. Note that generic style files (non-latex) often have a .tex extension.

input_content

used for cases when the file (or data) is plain TeX material that is expected to contribute content to the document (as opposed to pure definitions). A Mouth is opened onto the file, and subsequent reading and/or digestion will pull Tokens from that Mouth until it is exhausted, or closed.

input_definitions

TODO: Flesh out with the full infrastructure, incremental functionality for now.

install_tag

load_binding

loads a binding from the main binding dispatcher, if available+found

load_class

load_external_binding

loads a binding from an external binding dispatcher, if available+found

load_font_map

load_tex_content

merge_font

preload_font_map

process_options

require_package

This (and FindFile) needs to evolve a bit to support reading raw .sty (.def, etc) files from the standard texmf directories. Maybe even use kpsewhich itself (INSTEAD of pathname_find ???) Another potentially useful option might be that if we are reading a raw file, perhaps it should just get digested immediately, since it shouldn’t contribute any boxes.

require_resource

select_relaxng_schema

Selects the RelaxNG schema defining the XML output language