Module rtx_package::prelude::binding::counter::dialect

Counters

This is modelled on LaTeX’s counter mechanisms, but since it also provides support for ID’s, even where there is no visible reference number, it is defined in general. These id’s should be both unique, and parallel the visible reference numbers (as much as possible). Also, for consistency, we add id’s to unnumbered document elements (eg from \section*); this requires an additional counter (eg. UNsection) and mechanisms to track it.

Structs

BeginItemizeOptions

configuration for begin_itemize

NewCounterOptions

configuration for new_counter

Functions

add_to_counter

increments a named counter by a Number

begin_itemize

Prepare for an list (itemize/enumerate/description/etc) by determining the right counter (level) and binding the right \item ( $type@item, if $type is defined)

counter_value

Fetches the value associated with the counter C<$ctr>.

maybe_note_label

Use to note a discovered label to support label-derived refererence numbers Can by used by \label, among others. Note we only record the label if it hasn’t already been peeked, and consumed.

new_counter

Defines a new counter named $ctr. If within is defined, ctr will be reset whenever within is incremented.

ref_step_counter

Analog of \refstepcounter, steps the counter and returns a hash containing the keys refnum and id. This makes it suitable for use in a properties option to constructors. The id is generated in parallel with the reference number to assist debugging.

ref_step_id

For UN-numbered units. Like RefStepCounter, but only steps the “uncounter”, and returns only the id; This is useful for unnumbered cases of objects that normally get both a refnum and id.

ref_step_item_counter

Create id, and tags for an itemize type \item

rescue_caption_counters

Copies the current id, tags, and inlist counter values into whatsit properties

reset_counter

Resets the counter ctr to zero.

step_counter

Analog of \stepcounter, steps the counter and returns the expansion of \the$ctr Usually you should use ref_step_counter(ctr) instead.