rtx_package::prelude::binding::def
Module traits

A variety of traits helpful for auto-casting between the different components of the conversion toolchain

Traits

IntoBoolResult

Creates Result<bool> from some type T

IntoDigestedOptionResult

IntoDigestedResult

Creates a Result<Vec<Digested>> from some type T

IntoFontField

IntoOption

A trait for auto-wrapping a generic type T into Option<Y>, where Y can be inferred from context. (useful in macro helpers, such as NewDefaultV!)

IntoPropertiesResult

IntoRegisterValueOption

Creates an Option<RegisterValue> from some type T. Useful for Register getter closures

IntoResultArgWrap

Create a Result<ArgWrap> from any concrete type that Gullet may have a reader for. Used in auto-casting the data fetched by Parameter readers

IntoTokensResult

A trait for creating Result<Tokens> from all sensible concrete types one could return from e.g. a DefMacro closure