Module rtx_package::prelude::font

Structs

Font

This struct is a little interesting, as we want to pass overrides that partially modify (via a merge) the current font, in each definitional binding. To accommodate that with this struct, every single field needs to be an Option, in order to unambiguously tell the “intend” of override (Some) vs no intent (None).

Constants

MATH_FONTS

TEXT_FONTS

Statics

FONT_TEXT_DEFAULT

Functions

decode

Decode a codepoint using the fontmap for a given font and/or fontencoding. If encoding not provided, then lookup according to the current font’s encoding; the font family may also be used to choose the fontmap (think tt fonts!). When implicit is false, we are “explicitly” asking for a decoding, such as with \char, \mathchar, \symbol, DeclareTextSymbol and such cases. In such cases, only codepoints specifically within the map are covered; the rest are undef. If implicit is true, we’ll decode token content that has made it to the stomach: We’re going to assume that SOME sort of handling of input encoding is taking place, so that if anything above 128 comes in, it must already be Unicode!. The lower half plane still needs to go through decoding, though, to deal with TeX’s rearrangement of ASCII…

decode_fontname

decode_string

lookup_font_family

Global auxiliary for font family lookup

lookup_font_series

Global auxiliary for font series lookup

lookup_font_shape

Global auxiliary for font shape lookup

rationalize_font_size

Convert stanard font size names, such as tiny, Huge, etc to f64

relative_font_size

convert size to percent

Type Aliases

Fontmap