Module rtx_package::prelude::radix

“radix” may be a misnomer here. Primarily used to generate labels, or uniquifying suffixes to make ID’s simple radix conversion utilities

This module provides some simple utilities for radix conversion.

Functions

radix_alpha

converts the number into one or more lowercase latin letters

radix_format

(Internal) Converts the number into one of the char symbols

radix_format_str

(Internal) Converts the number into one of the str symbols

radix_greek

converts the number into one or more lowercase greek letters

radix_roman

converts the number as a lowercase roman numeral

radix_up_alpha

converts the number into one or more uppercase latin letters

radix_up_greek

converts the number into one or more uppercase greek letters

radix_up_roman

converts the number as a uppercase roman numeral