Module rtx_package::prelude::util

Auxilary utilities that do not participate in the main conversion abstraction

Modules

logger

log recording and reporting interface

pathname

helper methods for file system paths

radix

“radix” may be a misnomer here. Primarily used to generate labels, or uniquifying suffixes to make ID’s simple radix conversion utilities

text

helpers for extracting structured data from replacement (and other) strings