Crate rtx_package

Binding definitions for the LaTeXML converter, reimplemented in Rust

Re-exports

pub use package::*;

Modules

macros

package

Macros

AddToCounter

AddToMacro

AfterAssignment

AssignCatcode

AssignMapping

AssignMeaning

AssignValue

BeginItemize

BindInnerState

BindState

CounterValue

DeclareFontMap

DeclareOption

DefAccent

DefColumnType

DefConditional

DefConstructor

DefEnvironment

DefEnvironmentIWO

DefEnvironmentWO

DefKeyVal

Defines a new KeyVal Parameter in the given keyset, key and with optional prefix option.prefix. For descriptions of further parameters, see keyval::define.

DefLigature

DefMacro

A $prototype will be parsed into a command sequence and a list of parameters. Any macro arguments will be substituted for parameter indicators (eg #1) in the I or tokenized I and the result is used as the expansion of the control sequence. If I is used, it is called at expansion time and should return a list of tokens as its result.

DefMath

DefParameterType

DefParameterTypeWO

DefPrimitive

Define a primitive control sequence.

DefRegister

Defines a register with value as the initial value (a Number, Dimension, Glue, MuGlue or Tokens — I haven’t handled Box’s yet). Usually, the prototype is just the control sequence, but registers are also handled by prototypes like \count{Number}. DefRegister arranges that the register value can be accessed when a numeric, dimension, … value is being read, and also defines the control sequence for assignment.

Digest

DigestIf

DigestText

Dimension

DocType

Expand

Return $tokens with all tokens expanded

GetKeyVal

GetKeyVals

Glue

InnerPool

Loader shorthand for pool dependencies

InputDefinitions

InstallDefinition

Invocation

Invocation(); builds a representation of a command sequence invoked on its arguments

IsDefinable

IsDefined

IsDefinedToken

Let

LoadClass

LoadDefinitions

LoadFontMap

LoadPool

LookupBool

LookupCatcode

LookupColor

LookupDefinition

LookupFont

LookupMapping

LookupMappingKeys

LookupMeaning

LookupNumber

LookupRegister

LookupRegisterOrDefault

LookupString

LookupTokens

LookupValue

MergeFont

Merge the current font with the style specifications

MuGlue

NewCounter

NewCounterWO

PopValue

ProcessOptions

PushValue

RawTeX

RefStepCounter

convenience macro for api::counter_dialect::ref_step_counter

RefStepID

convenience macro for api::counter_dialect::ref_step_id

RegisterDocumentNamespace

RegisterNamespace

RelaxNGSchema

RemoveValue

RequirePackage

RequireResource

ResetCounter

Revert

SetPrefix

ShiftValue

StepCounter

T_ACTIVE

macro for an ACTIVE char token

T_ALIGN

macro for an ALIGN “&” token

T_ARG

macro for a numbered ARG token

T_BEGIN

macro for a BEGIN “{” token

T_COMMENT

macro for a COMMENT content token

T_CR

macro for a CR “\n” token

T_CS

macro for a command sequence token

T_LETTER

macro for a LETTER token

T_MARKER

macro for a tracing MARKER token

T_MATH

macro for a MATH “$” token

T_OTHER

macro for an OTHER code token

T_PARAM

macro for a PARAM “#” token

T_SPACE

macro for a SPACE token (default “ “)

T_SUB

macro for a SUB “_” token

T_SUPER

macro for a SUPER “^” token

Tag

Today

Tokenize

Tokenize($string); Tokenizes the string using the standard cattable, returning a LaTeXML::Core::Tokens

TokenizeInternal

TokenizeInternal($string); Tokenizes the string using the internal cattable, returning a LaTeXML::Core::Tokens

Tokens

convenience macro for assembling a Tokens object from different pieces (Token, Vec<Token>, Tokens)

TypedConditional

TypedMacro

TypedPrimitive

UnshiftValue

WithInnerState

XEquals

bind_state

bind_state_mut

compile_expansion

Macro for compiling string macro expansions into closures Approach borrowed from diesel-codegen

compile_prototype

Macro for compiling string literal prototypes into a Token and Parameters structs

compile_prototype_for_typed_conditional

Macro for compiling string binding prototypes into Conditional closures

compile_prototype_for_typed_macro

Macro for compiling string binding prototypes into Expandable closures Approach borrowed from diesel-codegen

compile_prototype_for_typed_primitive

Macro for compiling string binding prototypes into Primitive closures

compile_replacement

Macro for compiling string construction replacements into closures that execute the needed ops in libxml. Approach borrowed from diesel-codegen

compile_tokenize

Macro for compiling string literal tokens into their usual Tokens representation

compile_tokenize_internal

Macro for compiling string literal tokens into their preamble Tokens representation

defi_conditional

defi_constr

Internal auxiliary, only purpose is to bind state, then call the api::def_constructor function, where the interior construction logic resides.

defi_macro

Internal auxiliary, only purpose is to bind state, then call the api::def_macro function, where the interior macro installation logic resides.

defi_math

defi_opts

defi_primitive

defi_register

Internal auxiliary, only purpose is to bind state, then call the api::def_macro function, where the interior macro installation logic resides.

end_state_frame

load_model

parameter_rust_type

parse_prototype

A macro that uses rtx_codegen to expand a prototype at compile time, and initialize all parameters with the in-scope state at runtime. Returns a (cs, parameters) pair.

start_state_frame