Module rtx_package::engine::base_functions

Structs

KVSpec

Support for Key / Value arguments.

Functions

adjust_box_color

aligning_environment

and_split

c_revert

classify_box

cleanup_math

decode_math_char

dimension_to_spaces

a candidate for use by \hskip, \hspace, etc… ?

do_def

escapechar

in_svg

insert_block

This attempts to be a generalize vbox construction; The idea is to receeive block-like material, possibly wrapped in appropriate container which gets attributes. The contents are constructed in an ltx:CaptureBlock element, designed to accept all reasonable block material from several levels, and then determine which container element is most apprpriate for both the conent & context from block, logical-block or sectional-block, or the inline- variants.

keyvals_aux

lowercase_token

make_generic_message

neutralize_font

p_revert

parse_def_parameters

predigest_box_contents

Reading a Box’s content is crucially dependent on invoking the “{” token and obtaining a digested result Hence it is always needed to pair read_box_contents with its stomach-level counterpart, predigest_box_contents

reenter_text_mode

revert_spec

set_align_or_class

split_tokens

today

translate_attachment

Utility, not really TeX, but used by LaTeX, AmSTeX… Convert a vertical positioning, optional argument. t = “top”, b = “bottom”; default is “middle”. Note that the default for vattach attribute is “baseline”.

uppercase_token

writable_tokens

Converts $tokens to a string in the fashion of \message and others: doubles #, converts to string; optionally adds spaces after control sequences in the spirit of the B Book, “show_token_list” routine, in 292. [This could be a $tokens->unpackParameters, but for the curious space treatment]