Module rtx_package::package::binding::def::macros

Macros

AssignRegister

Font

build a Font from key=>val pairs

FontDirective

build a FontDirective from key=>val pairs (currently only FontDirective::Asset is supported in this macro)

NewDefault

given a struct $name, create a new instance of it using the given key=>val pairs and complete the remaining entries via the Default instance

NewDefaultV

Just like NewDefault, but adds a mandatory .into_option() to all values

Roman

Convert the number to upper case roman numerals, returning a list of LaTeXML::Core::Token

SetCounter

after_digest

after_digest_simple

after_digest_single

before_digest

before_digest_simple

before_digest_single

construct

count_unpack_opt

count_unpack_ref

count_unpack_to_string

forbidMath

getter

A closure for obtaining a RegisterValue, usually owned by a Register getter.

noreplacement

primitiveproc

create a PrimitiveClosure from the pieces, with a forced empty return value

prop_bool

prop_digested

prop_str

prop_string

prop_whatsit

properties

reader

reader_predigest

replacement

requireMath

reversion

reversion_digested

rewrite_replace_sub

roman

Convert the number to lower case roman numerals, returning a list of LaTeXML::Core::Token

setter

sizersub

tagsub

transfer_opt_default

Assumption: $receiver is HashMap<String,String>. If, and only if, $has_receiver does not have a value at slot $val, and $struct_source has a set value at $val, then transfers (with ownership) the $val field of a $struct_source into the $receiver.

undigested

unpack

unpack_opt

An alternative to unpack!(args=>arg1,arg2...) which allows for optional arguments.

unpack_opt_ref

Meant to be used for unpacking &Vec<Option> in particular.

unpack_to_number

unpack_to_string

unpacks a Vec<ArgWrap> argument into a flexary number of String-typed variables uses the usual .to_string() method from the Display trait.

unpack_to_token

unref

Convenience macro to flexibly unpack a collection of Vec<ArgWrap> arguments into individual Tokens variables.