Struct rtx_package::package::gullet::Gullet

pub struct Gullet {
    pub mouth: Option<MouthRuntime>,
    pub mouthstack: VecDeque<MouthRuntime, Global>,
    pub pending_comments: VecDeque<Token, Global>,
    /* private fields */
}

Fields
mouth: Option<MouthRuntime>mouthstack: VecDeque<MouthRuntime, Global>pending_comments: VecDeque<Token, Global>
Implementations

impl Gullet

pub fn flush_mouth(&mut self, state: &mut State)

This flushes a mouth so that it will be automatically closed, next time it’s read Corresponds to TeX’s \endinput

pub fn flush(&mut self, state: &mut State)

Obscure, but the only way I can think of to End!! (see \bye or \end{document}) Flush all sources (close all pending mouth’s)

pub fn has_more_input(&mut self) -> bool

pub fn open_mouth(&mut self, mouth: Mouth, autoclose: bool)

pub fn close_mouth( &mut self, forced: bool, state: &mut State ) -> Result<(), Error>

pub fn get_mouth(&self) -> Option<&Mouth>

pub fn get_mouth_mut(&mut self) -> Option<&mut Mouth>

pub fn mouth_is_open(&self, mouth: &Mouth) -> bool

pub fn read_token(&mut self, state: &mut State) -> Result<Option<Token>, Error>

pub fn read_x_token( &mut self, toplevel_opt: Option<bool>, commentsok: bool, state: &mut State ) -> Result<Option<Token>, Error>

Return the next unexpandable token from the input source, or None if there is no more input. If the next token is expandable, it is expanded, and its expansion is reinserted into the input. If commentsok, a comment read or pending will be returned.

pub fn read_raw_line(&mut self, state: &State) -> Option<String>

Read the next raw line (string); primarily to read from the Mouth, but keep any unread input!

pub fn unread(&mut self, tokens: Tokens)

Push the tokens back into the input stream to be re-read.

pub fn unread_mut(&mut self, tokens: &mut Tokens)

same as unread, but drains the tokens from its contents

pub fn unread_one(&mut self, token: Token)

same as unread, but only for a single Token

pub fn unread_vec(&mut self, tokens: Vec<Token, Global>)

same as unread, but does not require the Tokens wrapper

pub fn read_non_space( &mut self, state: &mut State ) -> Result<Option<Token>, Error>

Read a single non-space token

pub fn read_x_non_space( &mut self, state: &mut State ) -> Result<Option<Token>, Error>

Read a single expanded, non-space, token

pub fn read_balanced( &mut self, expanded: bool, state: &mut State ) -> Result<Option<Tokens>, Error>

Read a sequence of tokens balanced in {} assuming the { has already been read. Returns a Tokens list of the balanced sequence, omitting the closing }

pub fn read_keyword( &mut self, keywords: &[&str], state: &mut State ) -> Result<Option<String>, Error>

Match the input against a set of keywords; Similar to readMatch, but the keywords are strings, and Case and catcodes are ignored; additionally, leading spaces are skipped. AND, macros are expanded.

pub fn read_until( &mut self, delim: &Tokens, state: &mut State ) -> Result<Tokens, Error>

Return a (balanced) sequence tokens until a match against one of the Tokens in @delims. Note that Braces on input hides the contents from matching, so this assumes there wont be braces in $delim! But, see readUntilBrace for that case.

pub fn read_until_token( &mut self, t: Token, state: &mut State ) -> Result<Tokens, Error>

Convenience method wrapping around read_until TODO: This seems to be the wrong Rust type interface, we need to rework…

pub fn read_until_brace( &mut self, state: &mut State ) -> Result<Option<Tokens>, Error>

reads until it encounters a Catcode::BEGIN token

pub fn read_next_conditional( &mut self, state: &mut State ) -> Result<Option<(Token, ConditionalType)>, Error>

reads and discards tokens, until it encounters a conditional, if any

pub fn read_arg(&mut self, state: &mut State) -> Result<Tokens, Error>

Higher-level readers: Read various types of things from the input: tokens, non-expandable tokens, args, Numbers, …

pub fn read_optional( &mut self, default: Option<Tokens>, state: &mut State ) -> Result<Option<Tokens>, Error>

Read and return a LaTeX optional argument; returns C<$default> if there is no ‘[’, otherwise the contents of the []. Note that this returns an empty array if [] is present, i.e. “[contents]” in TeX will lead to Tokens(contents), otherwise returns None

pub fn if_next( &mut self, token: &Token, state: &mut State ) -> Result<bool, Error>

pub fn read_value( &mut self, value_type: RegisterType, state: &mut State ) -> Result<RegisterValue, Error>

pub fn read_register_value( &mut self, value_type: RegisterType, state: &mut State ) -> Result<Option<RegisterValue>, Error>

pub fn read_match( &mut self, choices: &[&Tokens], state: &mut State ) -> Result<Option<Tokens>, Error>

Match the input against one of the Token or Tokens in @choices; return the matching one or undef.

pub fn read_number(&mut self, state: &mut State) -> Result<Number, Error>

====================================================================== Integer, Number

= = | = |

pub fn read_normal_integer( &mut self, state: &mut State ) -> Result<Option<Number>, Error>

= | | ’ | “ | `

pub fn read_float(&mut self, state: &mut State) -> Result<Float, Error>

====================================================================== Float, a floating point number. Similar to factor, but does NOT accept comma! This is NOT part of TeX, but is convenient.

pub fn read_dimension(&mut self, state: &mut State) -> Result<Dimension, Error>

pub fn read_glue(&mut self, state: &mut State) -> Result<Glue, Error>

pub fn read_rubber( &mut self, mu: bool, state: &mut State ) -> Result<(Option<i64>, Option<FillCode>), Error>

pub fn read_mu_glue(&mut self, state: &mut State) -> Result<MuGlue, Error>

pub fn read_mu_dimension( &mut self, state: &mut State ) -> Result<MuDimension, Error>

pub fn read_mu_unit(&mut self, state: &mut State) -> Result<Option<i64>, Error>

pub fn read_tokens_value(&mut self, state: &mut State) -> Result<Tokens, Error>

Apparent behaviour of a token value (ie \toks#=)

pub fn skip_spaces(&mut self, state: &mut State) -> Result<(), Error>

pub fn skip_one_space(&mut self, state: &mut State) -> Result<(), Error>

pub fn setup_scan(&mut self)

pub fn reading_from_mouth<R, FnR>( &mut self, mouth: Mouth, state: &mut State, reader: FnR ) -> Result<R, Error>where FnR: FnOnce(&mut Gullet, &mut State) -> Result<R, Error>,

Do something, while reading stuff from a specific Mouth. This reads ONLY from that mouth (or any mouth openned by code in that source), and the mouth should end up empty afterwards, and only be closed here.

pub fn do_expand<T>( &mut self, tokens: T, outer_state: &mut State ) -> Result<Tokens, Error>where T: Into<Tokens>,