Struct rtx_package::package::ConstructorOptions

pub struct ConstructorOptions {

    pub nargs: Option<usize>,
    pub bounded: bool,
    pub mode: Option<String>,
    pub sizer: Option<Arc<dyn Fn(&Whatsit, &mut State) -> Result<(Dimension, Dimension, Dimension), Error> + 'static>>,
    pub before_digest: Vec<Arc<dyn Fn(&mut Stomach, &mut State) -> Result<Vec<Digested, Global>, Error> + 'static>, Global>,
    pub after_digest: Vec<Arc<dyn Fn(&mut Stomach, &mut Whatsit, &mut State) -> Result<Vec<Digested, Global>, Error> + 'static>, Global>,
    pub before_construct: Vec<Arc<dyn Fn(&mut Document, &Whatsit, &mut State) -> Result<(), Error> + 'static>, Global>,
    pub after_construct: Vec<Arc<dyn Fn(&mut Document, &Whatsit, &mut State) -> Result<(), Error> + 'static>, Global>,
    pub require_math: bool,
    pub forbid_math: bool,
    pub properties: Arc<dyn Fn(&mut Stomach, &Vec<Option<Digested>, Global>, &mut State) -> Result<HashMap<String, Stored, RandomState>, Error> + 'static>,
    pub capture_body: bool,
    pub font: Option<FontDirective>,
    pub after_digest_begin: Vec<Arc<dyn Fn(&mut Stomach, &mut Whatsit, &mut State) -> Result<Vec<Digested, Global>, Error> + 'static>, Global>,
    pub before_digest_end: Vec<Arc<dyn Fn(&mut Stomach, &mut State) -> Result<Vec<Digested, Global>, Error> + 'static>, Global>,
    pub after_digest_body: Vec<Arc<dyn Fn(&mut Stomach, &mut Whatsit, &mut State) -> Result<Vec<Digested, Global>, Error> + 'static>, Global>,
    pub reversion: Option<Reversion>,
    pub scope: Option<Scope>,
    pub robust: bool,
    pub locked: bool,
    pub alias: Option<String>,

}

configuration for creating a new Constructor

Fields
nargs: Option<usize>

number of arguments (if any)

bounded: bool

bouded mode (default: false)

mode: Option<String>

begin a named mode

sizer: Option<Arc<dyn Fn(&Whatsit, &mut State) -> Result<(Dimension, Dimension, Dimension), Error> + 'static>>

a SizingClosure to estimate the size of the digested box

before_digest: Vec<Arc<dyn Fn(&mut Stomach, &mut State) -> Result<Vec<Digested, Global>, Error> + 'static>, Global>

custom code to run immediately before the digestion phase

after_digest: Vec<Arc<dyn Fn(&mut Stomach, &mut Whatsit, &mut State) -> Result<Vec<Digested, Global>, Error> + 'static>, Global>

custom code to run immediately after the digestion phase

before_construct: Vec<Arc<dyn Fn(&mut Document, &Whatsit, &mut State) -> Result<(), Error> + 'static>, Global>

custom code to run immediately before the construction phase

after_construct: Vec<Arc<dyn Fn(&mut Document, &Whatsit, &mut State) -> Result<(), Error> + 'static>, Global>

custom code to run immediately after the construction phase

require_math: bool

requires to be used in math mode

forbid_math: bool

forbids use in math mode

properties: Arc<dyn Fn(&mut Stomach, &Vec<Option<Digested>, Global>, &mut State) -> Result<HashMap<String, Stored, RandomState>, Error> + 'static>

custom directives for computing box properties

capture_body: bool

should it capture the body as #body (default: false)

font: Option<FontDirective>

specify a font to use, or instructions how to compute which font to use

after_digest_begin: Vec<Arc<dyn Fn(&mut Stomach, &mut Whatsit, &mut State) -> Result<Vec<Digested, Global>, Error> + 'static>, Global>

custom code to run as digestion begins

before_digest_end: Vec<Arc<dyn Fn(&mut Stomach, &mut State) -> Result<Vec<Digested, Global>, Error> + 'static>, Global>

custom code to run just before digestion ends

after_digest_body: Vec<Arc<dyn Fn(&mut Stomach, &mut Whatsit, &mut State) -> Result<Vec<Digested, Global>, Error> + 'static>, Global>

custom code to run after #body has been digested

reversion: Option<Reversion>

provide tokens to revert to, or custom code for computing them

scope: Option<Scope>

Local/Global scope of installing this definition (default: Local)

robust: bool

is this a robust command sequence (default: false)

locked: bool

lock the definition for raw TeX overrides (default: false)

alias: Option<String>

alternative (command sequence) name, used for reversion