Module rtx_package::package::token

The Token and its Catcode

Macros

CharToken

Special case: a character needs swift string conversion, so let’s use a dedicated macro

Explode

Explode a string into a list of tokens, all w/catcode OTHER (except space).

ExplodeText

Similar to Explode, but convert letters to catcode LETTER and others to OTHER Hopefully, this is essentially correct WITHOUT resorting to catcode lookup?

T_ACTIVE

macro for an ACTIVE char token

T_ALIGN

macro for an ALIGN “&” token

T_ARG

macro for a numbered ARG token

T_BEGIN

macro for a BEGIN “{” token

T_COMMENT

macro for a COMMENT content token

T_CR

macro for a CR “\n” token

T_CS

macro for a command sequence token

T_END

macro for a new END “{” token

T_LETTER

macro for a LETTER token

T_MARKER

macro for a tracing MARKER token

T_MATH

macro for a MATH “$” token

T_OTHER

macro for an OTHER code token

T_PARAM

macro for a PARAM “#” token

T_RELAX

macro for T_CS(“\relax”)

T_SMUGGLE_THE

macro for a SmuggleThe token (see gullet::invoke_and_read_x_token)

T_SPACE

macro for a SPACE token (default “ “)

T_SUB

macro for a SUB “_” token

T_SUPER

macro for a SUPER “^” token

Token

Token constructor macro (defaults to OTHER code)

Structs

Token

The core immutable syntactic primitive resulting from TeX’s read-in and expansion process We allow the fields to be public, so that we can use builder macros such as

Enums

Catcode

A Token category code, as in TeX

Statics

TOKEN_ALIGN

constant for an ALIGN “&” token

TOKEN_BEGIN

constant for an END “}” token

TOKEN_CR

constant for a CR “\n” token

TOKEN_END

constant for a BEGIN “{” token

TOKEN_MATH

constant for a MATH “$” token

TOKEN_PARAM

constant for a PARAM “#” token

TOKEN_RELAX

constant for T_CS(“\relax”)

TOKEN_SPACE

constant for a SPACE “ “ token

TOKEN_SUB

constant for a SUB “_” token

TOKEN_SUPER

constant for a SUPER “^” token