Struct rtx_package::package::tokens::Tokens

pub struct Tokens(_);

Tokens are a thin wrapper over a vector of Token objects usually read from a Mouth. They are usually treated as an immutable interface, an have to be consumed via .unlist() for access to the underlying data.

Implementations

impl Tokens

pub fn new(tokens: Vec<Token, Global>) -> Tokens

Create a Tokens object from a Vec of individual Token

pub fn unlist(self) -> Vec<Token, Global>

Return a list of the tokens making up this Tokens

pub fn unlist_ref(&self) -> &Vec<Token, Global>

Return a reference to the tokens making up this Tokens

pub fn unlist_mut(&mut self) -> &mut Vec<Token, Global>

Return a mutable reference to the tokens making up this Tokens

pub fn is_empty(&self) -> bool

Are there any tokens at all contained in this Tokens object

pub fn len(&self) -> usize

Number of contained Token entries

pub fn revert(self) -> Vec<Token, Global>

Return a string containing the TeX form of the Tokens

pub fn to_number(&self) -> Number

to_number casts back to a parsed Number (usually via gullet.read_number) which had to be re-converted to a Tokens for reentering the expansion flow

pub fn to_dimension(&self) -> Dimension

to_dimension casts back to a parsed Dimension (usually via gullet.read_dimension) which had to be re-converted to a Tokens for reentering the expansion flow

pub fn to_glue(&self) -> Glue

to_glue casts back to a parsed Glue (usually via gullet.read_glue) which had to be re-converted to a Tokens for reentering the expansion flow

pub fn to_mu_glue(&self) -> MuGlue

to_mu_glue casts back to a parsed MuGlue (usually via gullet.read_mu_glue) which had to be re-converted to a Tokens for reentering the expansion flow

pub fn to_mu_dimension(&self) -> MuDimension

to_mu_dimension casts back to a parsed MuGlue (usually via gullet.read_mu_glue) which had to be re-converted to a Tokens for reentering the expansion flow

pub fn to_float(&self) -> Float

to_float casts back to a parsed Float (usually via gullet.read_float) which had to be re-converted to a Tokens for reentering the expansion flow

pub fn to_keyvals(&self, state: &State) -> KeyVals

to_keyvals casts back to a parsed KeyVals (usually via a KeyVals parameter type) which had to be re-converted to a Tokens for reentering the expansion flow

pub fn equals(&self, other: Tokens) -> bool

Methods for overloaded ops.

pub fn unwrap_or_default(self) -> Tokens

returns self, for compatibility convenience with Option

pub fn unwrap(&self) -> &Tokens

returns self, for compatibility convenience with Option

pub fn stringify(&self) -> String

A string form which is primarily used for error-reporting

pub fn be_digested( self, stomach: &mut Stomach, state: &mut State ) -> Result<Digested, Error>

digest the current Tokens

pub fn neutralize(self, extraspecials: &[char], state: &State) -> Tokens

Remove dont_expand, but preserve SMUGGLE_THE

pub fn is_balanced(&self) -> bool

Checks if any BEGIN/END code groups are correctly nested and closed

pub fn substitute_parameters(&self, args: &[Option<Cow<'_, Tokens>>]) -> Tokens

substitutes the parameters (ARG catcode) in a Tokens list for concrete arguments

pub fn without_dont_expand(self) -> Tokens

removes the smuggled token of all contained Token elements

pub fn untex(self) -> String

Consumes a Tokens to a string containing TeX that created it (or could have). Note that this is not necessarily the original TeX code; expansions or other substitutions may have taken place. Also note that the LaTeXML linebreak feature is always disabled here.)

pub fn pack_parameters(self) -> Tokens

Process the Catcode::PARAM tokens for use as a macro body (and other token lists)