Trait rtx_package::package::BoxOps

pub trait BoxOps: Object {

    fn be_absorbed(
        &self,
        document: &mut Document,
        state: &mut State
    ) -> Result<Vec<Node, Global>, Error>;
    fn get_string(&self, state: &State) -> Result<Cow<'_, str>, Error>;
    fn get_properties(
        &self
    ) -> &HashMap<String, Stored, BuildHasherDefault<FxHasher>>;
    fn set_property<T>(&mut self, key: &str, value: T)
    where
        T: Into<Stored>;
    fn has_property(&self, key: &str) -> bool;
    fn get_property_bool(&self, _key: &str) -> bool;
    fn get_font(&self, state: &mut State) -> Result<Option<Cow<'_, Font>>, Error>;
    fn compute_size(
        &self,
        options: HashMap<String, Stored, BuildHasherDefault<FxHasher>>,
        state: &mut State
    ) -> Result<(Dimension, Dimension, Dimension), Error>;

    fn unlist(&self) -> Vec<Digested, Global>  { ... }
    fn get_tokens(&self) -> Option<&Tokens> { ... }
    fn get_property(&self, key: &str) -> Option<Cow<'_, Stored>> { ... }
    fn get_body(&self) -> Option<Digested> { ... }
    fn set_font(&mut self, _font: Arc<Font>) { ... }
    fn set_width<T>(&mut self, width: T)
    where
        T: Into<Stored>,
    { ... }
    fn get_width(
        &mut self,
        options: Option<HashMap<String, Stored, BuildHasherDefault<FxHasher>>>,
        state: &mut State
    ) -> Result<Option<RegisterValue>, Error> { ... }
    fn set_height<T>(&mut self, width: T)
    where
        T: Into<Stored>,
    { ... }
    fn get_height(&self, _state: &State) -> Option<RegisterValue> { ... }
    fn set_depth<T>(&mut self, width: T)
    where
        T: Into<Stored>,
    { ... }
    fn get_depth(&self, _state: &State) -> Option<RegisterValue> { ... }
    fn get_size(
        &self,
        options: Option<HashMap<String, Stored, BuildHasherDefault<FxHasher>>>,
        state: &mut State
    ) -> Result<(Dimension, Dimension, Dimension), Error> { ... }
    fn compute_size_store(
        &mut self,
        options: HashMap<String, Stored, BuildHasherDefault<FxHasher>>,
        state: &mut State
    ) -> Result<(), Error> { ... }

}

Common operations for Box-like (digested) data

Required Methods

fn be_absorbed(
&self,
document: &mut Document,
state: &mut State
) -> Result<Vec<Node, Global>, Error>

absorb the current object into the Document XML - returning the corresponding nodes

fn get_string(&self, state: &State) -> Result<Cow<'_, str>, Error>

build a string representation of the underlying digested data

fn get_properties(
&self
) -> &HashMap<String, Stored, BuildHasherDefault<FxHasher>>

get the map of named properties

fn set_property<T>(&mut self, key: &str, value: T)where
T: Into<Stored>,

set a named property (allows all Stored types for values)

fn has_property(&self, key: &str) -> bool

checks if a property key has been set

fn get_property_bool(&self, _key: &str) -> bool

obtains a boolean property value (false unless Stored::Bool)

fn get_font(&self, state: &mut State) -> Result<Option<Cow<'_, Font>>, Error>

gets the associated font, if any

fn compute_size(
&self,
options: HashMap<String, Stored, BuildHasherDefault<FxHasher>>,
state: &mut State
) -> Result<(Dimension, Dimension, Dimension), Error>

computes and returns the size of a box-like object

Provided Methods

fn unlist(&self) -> Vec<Digested, Global>

If composite, unwrap into the contained digested objects (or return self)

fn get_tokens(&self) -> Option<&Tokens>

get the underlying tokens (preceding digestion)

fn get_property(&self, key: &str) -> Option<Cow<'_, Stored>>

get a single named property

fn get_body(&self) -> Option<Digested>

obtains the “body” of a digested object which captured it

fn set_font(&mut self, _font: Arc<Font>)

sets an associated font

fn set_width<T>(&mut self, width: T)where
T: Into<Stored>,

sets a “width” property, for sizing

fn get_width(
&mut self,
options: Option<HashMap<String, Stored, BuildHasherDefault<FxHasher>>>,
state: &mut State
) -> Result<Option<RegisterValue>, Error>

gets the “width” property value, if any

fn set_height<T>(&mut self, width: T)where
T: Into<Stored>,

sets a “height” property value, for sizing

fn get_height(&self, _state: &State) -> Option<RegisterValue>

gets the “height” property value, if any

fn set_depth<T>(&mut self, width: T)where
T: Into<Stored>,

sets a “depth” property value, for sizing

fn get_depth(&self, _state: &State) -> Option<RegisterValue>

gets the “depth” property value, if any

fn get_size(
&self,
options: Option<HashMap<String, Stored, BuildHasherDefault<FxHasher>>>,
state: &mut State
) -> Result<(Dimension, Dimension, Dimension), Error>

gets the box size as a triple of (width, height, depth)

fn compute_size_store(
&mut self,
options: HashMap<String, Stored, BuildHasherDefault<FxHasher>>,
state: &mut State
) -> Result<(), Error>

deprecated/to be revisited - computes and caches the size of a box-like object