Module rtx_package::prelude

Generic support layer for any LaTeXML binding file

Re-exports

pub use log;

pub use crate::engine::base_functions::*;

pub use crate::engine::latex_functions::*;

pub use crate::package::*;

Modules

alignment

Support for TeX-like Alignments

arena

An arena for interning strings (global, mutable, single thread only)

aux_macros

Auxiliary macros Helper macros for quicker and more idiomatic construction and access to data structures.

binding

The programmable API foundation for creating bindings of LaTeX sty/cls libraries

comment

TeX comments as standalone objects

common

Common abstractions that are useful at various stages of Core processing

data

definition

All possible definitions for TeX-native commands (expandable, primitive, constructor,…)

digested

a shared interface for digested objects Interface layer for the full range of digested objects

document

An abstraction layer over the converted XML document

font

gullet

The Gullet is responsible for reading Tokens and other data from the Mouth

keyval

A LaTeX-like Key-Value object Key-Value Definitions

keyvals

A collection of Key-Value objects, typically from a single LaTeX argument

ligature

Rules for combining together characters and other text rules

list

A list of Digested objects

model

mouth

The mouth is a thin interface over a file, responsible for reading characters and associating them with catcodes

parameter

The abstraction layer used by the Gullet to read arguments for the various kinds of TeX object definitions

pathname

helper methods for file system paths

radix

“radix” may be a misnomer here. Primarily used to generate labels, or uniquifying suffixes to make ID’s simple radix conversion utilities

rewrite

Rules for rewriting the constructed XML document, after core processing has completed

setup_binding_language

state

A global, singleton, mutable state - hosts almost all TeX-facing runtime information for the conversion

stomach

The stomach is an abstraction responsible for digesting Tokens and Registers prepared by the Gullet into Boxes

tbox

A TeX-like digested Box

token

The Token and its Catcode

tokens

Grouping Tokens together Token List constructors.

util

Auxilary utilities that do not participate in the main conversion abstraction

whatsit

A TeX-like digested Whatsit

Macros

AssignRegister

CharToken

Special case: a character needs swift string conversion, so let’s use a dedicated macro

Debug

Error

Explode

Explode a string into a list of tokens, all w/catcode OTHER (except space).

ExplodeChars

ExplodeText

Similar to Explode, but convert letters to catcode LETTER and others to OTHER Hopefully, this is essentially correct WITHOUT resorting to catcode lookup?

Fatal

Font

build a Font from key=>val pairs

FontDirective

build a FontDirective from key=>val pairs (currently only FontDirective::Asset is supported in this macro)

Info

NewDefault

given a struct $name, create a new instance of it using the given key=>val pairs and complete the remaining entries via the Default instance

NewDefaultV

Just like NewDefault, but adds a mandatory .into_option() to all values

Number

Roman

Convert the number to upper case roman numerals, returning a list of LaTeXML::Core::Token

SetCounter

T_ACTIVE

macro for an ACTIVE char token

T_ALIGN

macro for an ALIGN “&” token

T_ARG

macro for a numbered ARG token

T_BEGIN

macro for a BEGIN “{” token

T_COMMENT

macro for a COMMENT content token

T_CR

macro for a CR “\n” token

T_CS

macro for a command sequence token

T_END

macro for a new END “{” token

T_LETTER

macro for a LETTER token

T_MARKER

macro for a tracing MARKER token

T_MATH

macro for a MATH “$” token

T_OTHER

macro for an OTHER code token

T_OTHER_CHAR

T_OTHER from a single character

T_PARAM

macro for a PARAM “#” token

T_RELAX

macro for T_CS(“\relax”)

T_SPACE

macro for a SPACE token (default “ “)

T_SUB

macro for a SUB “_” token

T_SUPER

macro for a SUPER “^” token

Token

Token constructor macro (defaults to OTHER code)

Tokens

convenience macro for assembling a Tokens object from different pieces (Token, Vec<Token>, Tokens)

Warn

after_digest

after_digest_simple

after_digest_single

before_digest

before_digest_simple

before_digest_single

construct

count_unpack_ref

fatal

fontmap

Instantiates a Font, using the Font fields as keys, and calling .into() for each value. The specification can be partial - missing fields are taken via the Default trait.

forbidMath

generate_message

getter

A closure for obtaining a RegisterValue, usually owned by a Register getter.

map

A flexary macro for constructing HashMap<String, T> maps, where T is generic.

mixrc

A variant on mixvec! where each argument receives an additional .into() call best used with an outer context that explicitly provides the expected type, such as

mixvec

A variant on vec! where each argument receives an additional .into() call best used with an outer context that explicitly provides the expected type, such as

noreplacement

predigest

print_stderr

prints a to STDERR without a line break

println_stderr

prints a single line to STDERR

prop_bool

prop_digested

prop_str

prop_string

prop_whatsit

properties

raw_char_map

A flexary macro for constructing HashMap<char, T> maps, where T is generic

raw_map

A flexary macro for constructing HashMap<K, V> maps, where K and V are both generic (inferred at time of use)

reader

replacement

report

report_mut

requireMath

reversion

reversion_digested

rewrite_replace_sub

roman

Convert the number to lower case roman numerals, returning a list of LaTeXML::Core::Token

s

The s! macro is a briefer alias for format!

set

Simple generic helper for HashSet<K,V> creation Source: https://riptutorial.com/rust/example/4149/create-a-hashset-macro

setter

sizersub

some

The some! macro transforms data in type S to Option<Into<T>> (always wrapping with Some)

static_map

A flexary macro for constructing HashMap<&'static str, &'static str> maps

stored_map

A flexary macro for constructing SymHashMap<Stored> maps

string_keys_map

A flexary macro for constructing HashMap<String, T> maps

string_map

A flexary macro for constructing HashMap<String, String> maps

sym_map

A flexary macro for constructing SymHashMap<T> maps

tagsub

transfer_opt_default

Assumption: $receiver is HashMap<String,String>. If, and only if, $has_receiver does not have a value at slot $val, and $struct_source has a set value at $val, then transfers (with ownership) the $val field of a $struct_source into the $receiver.

unpack_opt

Try to efficiently unwrap a Vec<T> into a [T;n] for $arg1…$argn

unpack_opt_ref

Try to efficiently unwrap a &Vec<T> into a &[T;n] for $arg1…$argn

unported

unref

Convenience macro to flexibly unpack a collection of Vec<ArgWrap> arguments into individual Tokens variables.

Structs

Alignment

AlignmentConfig

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

BeginItemizeOptions

configuration for begin_itemize

Cell

Conditional

A Conditional definition; Expandable.

ConditionalOptions

configurations for a conditional.

ConstructorOptions

configuration for creating a new Constructor

Core

The Core conversion runtime

CoreOptions

Configuration for the Core processing

Digested

An Rc-guarded abstraction for any object encountered at the “digested” phase of processing

Dimension

Document

Error

Expandable

ExpandableOptions

FindFileOptions

configuration for searching for a file in the local filesystem

Float

Font

This struct is a little interesting, as we want to pass overrides that partially modify (via a merge) the current font, in each definitional binding. To accommodate that with this struct, every single field needs to be an Option, in order to unambiguously tell the “intend” of override (Some) vs no intent (None).

Glue

Gullet

InputDefinitionOptions

a configuration for loading LaTeX definition files (such as .sty, .cls, and their bindings)

InputOptions

configuration for input of a TeX source (content files mostly)

KeyVals

KeyvalConfig

disable a given key-val Configuration fields for declaring a new KeyVal pattern

KeyvalsConfig

Lazy

A value which is initialized on the first access.

Ligature

List

Lists can contain any Digested items, such as boxes, whatsits or other lists

Localized

These are fields realized via Perl’s “local” mechanism in LaTeXML, but (for now) require explicit “expire” calls in Rust. Ideally their ergonomics gets improved, or we gradually phase them out.

Locator

LogState

MathLigatureOptions

MathPrimitive

MathPrimitiveOptions

Mouth

MouthOptions

MouthRuntime

MuDimension

MuGlue

Namespace

An xml namespace

NewCounterOptions

configuration for new_counter

Node

An xml node

Number

Parameter

Parameters

Primitive

PrimitiveOptions

Rc

A single-threaded reference-counting pointer. ‘Rc’ stands for ‘Reference Counted’.

Regex

A compiled regular expression for searching Unicode haystacks.

Register

A struct representing a TeX register

RegisterOptions

configuration for creating a new Register

RequireOptions

Resource

Rewrite

RewriteOptions

State

The state efficiently maintain the bindings in a TeX-like fashion. bindings associate data with keys (eg definitions with macro names) and respect TeX grouping; that is, an assignment is only in effect until the current group (opened by \bgroup) is closed (by \egroup).

StateOptions

state fields allowed for customization during construction

Stomach

The Stomach is responsible for digesting tokens into boxes, lists, etc.

SymHashMap

A convenience abstraction over a String-keyed HashMap typically used for HashMap<String,Stored> states. The goal is to support both a string interface, as well as the interned tickets interface, while avoiding String allocations internally.

TagOptions

Tbox

Box is a Rust keyword, so we use “Tbox” instead, as in “TeX Box”

Template

Token

The core immutable syntactic primitive resulting from TeX’s read-in and expansion process We allow the fields to be public, so that we can use builder macros such as

Tokens

Tokens are a thin wrapper over a vector of Token objects usually read from a Mouth. They are usually treated as an immutable interface, an have to be consumed via .unlist() for access to the underlying data.

UndoFrame

For each of several tables (being “value”, “meaning”, “catcode” or other space of names), each table maintains the bound values, and “undo” defines the stack frames

VecDeque

A double-ended queue implemented with a growable ring buffer.

Whatsit

Represents a digested object that can generate arbitrary elements in the XML Document.

Enums

Align

ArgWrap

Catcode

A Token category code, as in TeX

Catcodes

High-level catcode profiles

ConditionalType

classify the standard pieces of a conditional

Cow

A clone-on-write smart pointer.

DigestedData

These are all kinds of data which we consider officially supported as outputs from the digestion phase of TeX, i.e. from invoking a token.

ErrorCategory

ErrorTarget

ExpansionBody

FontDirective

LogStatus

NodeType

Types of xml nodes

PrimitiveBody

RegisterType

The type of a TeX Register

RegisterValue

The values that can be read by, and stored in, a Register

Reversion

Scope

installation scope in the state_tables

Stored

The original global state (in Perl) allowed arbitrary values. To stay consistent, we create an extremely permissive struct that affords all essential kinds of values that appear essential.

TableName

the kinds of tables bookkept in the State

TagOptionName

TexMode

The current TeX processing mode

Constants

NO_TOKENS

Use this to avoid reallocating a new empty Vec each time you need a placeholder Tokens return value

UNITY

UNITY_F64

XML_NS

Statics

ANY_SYM

the unique symbol for str value “ANY”

ARG_HOLE

regex for the usual argument placeholders #1-#9

CAPTURE_SYM

the unique symbol for str value “ltx:Capture”

CONDITIONAL_CS_RE

regex for the prefix of a conditional command sequence

DIRTY_ID_IDIOM_RE

regex for ${}^{label}$

DONT_EXPAND_SYM

the unique symbol for str value “\dont_expand”

DTD_SYM

the unique symbol for str value “DTD”

EMPTY_SYM

the unique symbol for the empty str value “”

FONT_SYM

the unique symbol for str value “font”

GLOBAL_DEFS_SYM

the unique symbol for str value “\globaldefs”

GULLET

HAS_ARG_OR_CS

regex for a TeX argument specifier or any command sequence

H_COMMENT_SYM

the unique symbol for str value “#COMMENT”

H_DEFAULT_SYM

the unique symbol for str value “#default”

H_DOC_SYM

the unique symbol for str value “#Document”

H_DTD_SYM

the unique symbol for str value “#DTD”

H_PCDATA_SYM

the unique symbol for str value “#PCDATA”

H_PI_SYM

the unique symbol for str value “#ProcessingInstruction”

LEADING_PROTOCOL_RE

regex for the prefix of a protocol (such as literal:)

LTX_P_SYM

the unique symbol for str value “ltx:p”

LTX_STAR_SYM

the unique symbol for str value “ltx:*”

MATH_SYM

the unique symbol for str value “math”

NON_ID_CHARSET_RE

regex for characters not expected in a usual id attribute

NO_PROPERTIES

RELAXNG_SYM

the unique symbol for str value “RelaxNG”

REPORT

SPACES_RE

regex for one-or-more spaces

STOMACH

TEXT_SYM

the unique symbol for str value “text”

TILDE_NOISE_RE

regex for a strange noisy TeX \\~{}

TOKEN_ALIGN

constant for an ALIGN “&” token

TOKEN_BEGIN

constant for a BEGIN “{” token

TOKEN_CR

constant for a CR “\n” token

TOKEN_END

constant for an END “}” token

TOKEN_EXPANDAFTER

constant for T_CS(“\expandafter”)

TOKEN_MATH

constant for a MATH “$” token

TOKEN_PARAM

constant for a PARAM “#” token

TOKEN_RELAX

constant for T_CS(“\relax”)

TOKEN_SPACE

constant for a SPACE “ “ token

TOKEN_SUB

constant for a SUB “_” token

TOKEN_SUPER

constant for a SUPER “^” token

TRAILING_SLASH_RE

regex for a trailing slash (trivial, but aids replacement of said slash)

UNITS

Used in conversion to scaled points

WILD_CARD_SYM

the unique symbol for str value “WildCard”

XML_ID_SYM

the unique symbol for str value “xml:id”

Traits

BoxOps

Common operations for Box-like (digested) data

Definition

FromStr

Parse a value from a string

IntoBoolResult

Creates Result<bool> from some type T

IntoDigestedOptionResult

IntoDigestedResult

Creates a Result<Vec<Digested>> from some type T

IntoFontField

IntoOption

A trait for auto-wrapping a generic type T into Option<Y>, where Y can be inferred from context. (useful in macro helpers, such as NewDefaultV!)

IntoPropertiesResult

IntoRegisterValueOption

Creates an Option<RegisterValue> from some type T. Useful for Register getter closures

IntoResultArgWrap

Create a Result<ArgWrap> from any concrete type that Gullet may have a reader for. Used in auto-casting the data fetched by Parameter readers

IntoTokensResult

A trait for creating Result<Tokens> from all sensible concrete types one could return from e.g. a DefMacro closure

NumericOps

Object

Base object for all LaTeXML Objects.

Functions

activate_scope

Activates all stashed definitions for the named scope. No-op if the scope is already active.

add_search_path

add_to_counter

increments a named counter by a Number

after_assignment

run the accumulated directives from \afterassignment

align_group_count

allocate_register

assign_alignment

assign_catcode

assigns a Catcode for a given character

assign_delcode

like assign_catcode but targets Delcode and its table

assign_font

a convenience method to globally asign a Font to the “font” key

assign_lccode

like assign_catcode but targets Lccode and its table

assign_mapping

assign_mathcode

like assign_catcode but targets Mathcode and its table

assign_meaning

$meaning should be a definition (for defining active control sequences) or another token, for \let

assign_register

assign_sfcode

like assign_catcode but targets Sfcode and its table

assign_uccode

like assign_catcode but targets Uccode and its table

assign_value

assigns a Stored value at the given key and scope

assign_value_sym

assigns a Stored value at the given (arena ticket!) key and scope

begin_itemize

Prepare for an list (itemize/enumerate/description/etc) by determining the right counter (level) and binding the right \item ( $type@item, if $type is defined)

begin_mode

Begin processing in mode; one of “text”, “display-math” or “inline-math”. This also begins a new level of grouping and switches to a font appropriate for the mode.

begin_semiverbatim

begins a semiverbatim frame, neutralizing the usual + requested characters

begingroup

Begin a new level of binding by pushing a new stack frame.

bgroup

Begin a new level of binding by pushing a new stack frame, and a new level of boxing the digested output.

build_invocation

Creates a single Tokens representing a TeX invocation of the lead token over a list of arguments.

checkin_value

Returns a value into its Stored::None placeholder (see checkout_value for taking it)

checkout_value

Replaces the value in question with Stored::None (see checkin_value for returning it)

clean_bib_key

cleans a string down to characters acceptable for a bibliography key

clean_id

cleans a string down to characters acceptable for an id attribute

clean_label

cleans a string down to characters acceptable for a label attribute

clean_url

cleans a string down to characters acceptable for a URL

clear_prefixes

clears the global prefixes

clone_box_list

close_mouth

coerce_cs

unconditionally wraps a CS token around a string

compose_url

builds a complete url from fragments

compute_indirect_model

The indirect model includes all elements allowed as direct children, and all descendents of a node that can be inserted after autoOpen’ing intermediate elements. This model therefor includes information from the Schema, as well as auto_open information that may be introduced in binding files.

convert_latex_args

Convert a LaTeX-style argument spec to our Package form. Ie. given $nargs and $optional, being the two optional arguments to something like \newcommand, convert it to the form we use

convert_unit

convert a unit name into a f64 scaling factor over sp

counter_value

Fetches the value associated with the counter C<$ctr>.

current_frame_message

explain the current frame

current_verbosity

deactivate_scope

Removes any definitions that were associated with the named scope. Normally not needed, since a scopes definitions are locally bound anyway.

decrement_align_group_count

def_conditional

Define a conditional control sequence. Its processing takes place in the Gullet. The test is applied to the arguments (if any), which determines which branch is executed. If the test is undefined, the conditional is a “user defined” one; Two additional primitives are defined \footrue and \foofalse; the test is then determined by the most recently called of those.

def_constructor

The Constructor is where LaTeXML really starts getting interesting; invoking the control sequence will generate an arbitrary XML fragment in the document tree. More specifically: during digestion, the arguments will be read and digested, creating a Whatsit to represent the object. During absorption by the Document, the Whatsit will generate the XML fragment according to the compiled_replacement.

def_environment

Defines an Environment that generates a specific XML fragment. compiled_replacement is of the same form as for DefConstructor, but will generally include reference to the #body property. Upon encountering a \begin{env}: the mode is switched, if needed, else a new group is opened; then the environment name is noted; the beforeDigest hook is run. Then the Whatsit representing the begin command (but ultimately the whole environment) is created and the after_digest_begin hook is run. Next, the body will be digested and collected until the balancing \end{env}. Then, any after_digest hook is run, the environment is ended, finally the mode is ended or the group is closed. The body and \end{env} whatsit are added to the \begin{env}’s whatsit as body and trailer, respectively.

def_macro

Defines the macro expansion for a cs: a macro control sequence that reads parameters as specified by paramlist and is expanded during macro expansion time in the Gullet. See ExpansionBody for the possible kinds of expansion material.

def_math

Define a Mathematical symbol or function. There are two sets of cases: (1) If the presentation appears to be TeX code, we create an XMDual, since the presentation may end up with structure, etc. (2) But if the presentation is a simple string, or unicode, it is just the content of the symbol; even if the function takes arguments.

def_math_constructor

Uses of DefMath without arguments, but with constructor-like options, are realized via a Constructor definition

def_math_dual

Advanced math replacements require a XMDual representation

def_math_primitive

EXPERIMENT: Introduce an intermediate case for simple symbols Define a primitive that will create a Tbox with the appropriate set of XMTok attributes.

def_primitive

Defines a primitive control sequence; a primitive is processed during digestion (in the Stomach), after macro expansion but before Construction time. Primitive control sequences generate Boxes or Lists, generally containing basic Unicode content, rather than structured XML. Primitive control sequences are also executed for side effect during digestion, effecting changes to the State.

def_register

Defines a register with value as the initial value (a Number, Dimension, Glue, MuGlue or Tokens — I haven’t handled Box’s yet). Usually, the prototype is just the control sequence, but registers are also handled by prototypes like \count{Number}. DefRegister arranges that the register value can be accessed when a numeric, dimension, … value is being read, and also defines the control sequence for assignment.

digest

Digest a list of tokens independent from any current Gullet. Typically used to digest arguments to primitives or constructors. Returns a List containing the digested material.

digest_if

digest_literal

digest_next_body

Return the digested List after reading and digesting a body from the its Gullet. The body extends until the current level of boxing or environment is closed.

digest_text

do_expand

dualize_arglist

Given a list of Tokens (to be expanded into mathematical objects) return two lists: (1) The Tokens’ wrapped in an XMAarg, with an ID added (2) a corresponding list of Tokens creating XMRef’s to those IDs

egroup

End a level of binding by popping the last stack frame, undoing whatever bindings appeared there, and also decrementing the level of boxing.

end_mode

End processing in mode; an error is signalled if stomach is not currently in mode. This also ends a level of grouping.

end_semiverbatim

end by just calling pop_frame

endgroup

End a level of binding by popping the last stack frame, undoing whatever bindings appeared there.

ensure_tag_property

expire_align_group_count

expire_current_token

expires the most recent (localized) current token.

expire_dual_branch

expire (localized) flag for “dual branch”

expire_ifframe

expires the most recent (originally Perl-local) IfFrame

expire_local_box_list

expire_reading_alignment

expire_state_unlocked

extend_box_list

find_file

search for a file as prescribed by a FindFileOptions configuration

floatformat

Utility for formatting sane numbers.

flush

Obscure, but the only way I can think of to End!! (see \bye or \end{document}) Flush all sources (close all pending mouth’s)

flush_mouth

This flushes a mouth so that it will be automatically closed, next time it’s read Corresponds to TeX’s \endinput

generate_error_stub

Generate a stub definition for an undefined control-sequence, along with appropriate error messge.

generate_ligature_id

simple id generator for a ligature

get_active_scopes

return the currently activated named scopes

get_bindings_dispatch

get_boxing_level

get the current boxing level

get_current_token

gets the (localized) current token

get_dual_branch

get the current value for “dual branch”

get_extra_bindings_dispatch

get_frame_depth

Determine depth of group nesting created by {,},\bgroup,\egroup,\begingroup,\endgroup by counting all frames which are not Daemon frames (and thus don’t possess FRAME_LOCK). This may give incorrect results for some special environments (e.g. minipage)

get_graphics_paths

get_ifframe

retrieves the most recent (originally Perl-local) IfFrame

get_indirect_model_relationship

get_input_encoding

get_known_scopes

return all known named scopes

get_location

Comment-oriented location string, based on get_locator

get_locator

Get the current location of input getting read

get_nomathparse_flag

get_pending_comment

get_prefix

gets the current value of a named prefix

get_reading_alignment

get_script_level

ScriptLevel is similar to boxing level, but relative to current Math mode’s level This is used for the scriptpos attribute to recognize overlapping sccripts. Making it relative to the math’s level avoids unnecessary changes

get_search_paths

get_status

get_tag_property

get_xmarg_id

graphics_paths_push_front

has_indirect_model

has_meaning

has_more_input

Check if there is more input to be read from the current mouth

has_reading_alignment

has_search_paths

has_value

Check if the Value table contains a given key

if_next

increment_align_group_count

initialize_gullet

Initialize (or reset, if reentrant) a Gullet to its default empty state

initialize_report

initialize_stomach

Initialize various stomach parameters, preload, etc.

input

This is essentially the \input equivalent; we are most likely expecting to get actual content, (possibly with definitions included, as well) but might actually be getting pure definitions, (like a proper style file) in which case we may really want to load a binding. Note that generic style files (non-latex) often have a .tex extension.

input_content

used for cases when the file (or data) is plain TeX material that is expected to contribute content to the document (as opposed to pure definitions). A Mouth is opened onto the file, and subsequent reading and/or digestion will pull Tokens from that Mouth until it is exhausted, or closed.

input_definitions

TODO: Flesh out with the full infrastructure, incremental functionality for now.

install_definition

A shorthand for installing definitions

install_tag

into_pin

invoke_token

Invoke a token; If it is a primitive or constructor, the definition will be invoked, possibly arguments will be parsed from the Gullet. Otherwise, the token is simply digested: turned into an appropriate box. Returns a list of boxes/whatsits.

is_column_end

is_definable

Check if the token is not yet defined, or let to \relax

is_defined

Is defined in the LaTeX-y sense of also not being let to \relax.

is_defined_token

Token variant of is_defined. Defined in the LaTeX-y sense of also not being let to \relax.

is_dont_expandable

Whether token is affected by \noexpand

is_value_bound

Was name bound? If frame is given, check only whether it is bound in that frame (0 is the topmost).

join

len

let_i

Let macro setter

load_binding

loads a binding from the main binding dispatcher, if available+found

load_class

load_external_binding

loads a binding from an external binding dispatcher, if available+found

load_font_map

load_tex_content

local_align_group_count

local_build_template

local_current_token

localizes a new current token. see Stomach::invoke_token

local_reading_alignment

local_state_unlocked

lookup_alignment

lookup_bool

A bit of Perl “existence as truth” semantics mixed in with proper boolean lookup

lookup_catcode

Lookup & assign a character’s Catcode

lookup_conditional

lookup_definition

used for expansion & various queries Since we’re not doing digestion here, we don’t need to handle mathactive, nor cs let to executable tokens This returns a definition object, or undef

lookup_definition_stored

Returns a definition as Stored so that one can call .read_arguments, which can’t be specialized during compile-time over a trait object Instead we’ll dispatch via Stored at runtime, to allow generic calls

lookup_delcode

like lookup_catcode but targets Delcode and its table

lookup_digestable_definition

Recognizes mathactive tokens in math mode and also looks for cs that have been let to other `executable’ tokens. Returns a definition object, or a “self inserting” token. Used for digestion.

lookup_dimension

a variant of lookup_value that casts the value into Dimension

lookup_expandable

lookup_font

convenience method to lookup the current value at the “font” key

lookup_glue

a variant of lookup_value that only recognizes a Stored::Glue

lookup_int

like lookup_value but only recognizes Int, Bool and Number variants of Stored (default: 0)

lookup_lccode

like lookup_catcode but targets Lccode and its table

lookup_mapping

lookup_mathcode

like lookup_catcode but targets Mathcode and its table

lookup_mathcode_sym

lookup_mathfont

convenience method to lookup the current value at the “mathfont” key

lookup_meaning

Get the `Meaning’ of a token. For active control sequences this may give the definition object (if defined) or another token (if \let) or undef Any other token is returned as is.

lookup_muglue

a variant of lookup_value that only recognizes a Stored::Glue

lookup_number

a variant of lookup_value that casts the value into Number

lookup_register

lookup_register_definition

A specialized version of lookup_definition for registers, since we can’t adequately perform multi-dispatch when we have a “Self: Sized” for the Definition trait object.

lookup_sfcode

like lookup_catcode but targets Sfcode and its table

lookup_string

like lookup_value, but casts the entry into a String (empty if None)

lookup_string_sym

like lookup_value, but casts the entry into a SymStr from the string interner (EMPTY_SYM if None)

lookup_token

a variant of lookup_value that only recognizes a Stored::Token

lookup_tokens

a variant of lookup_value that casts the response into Tokens

lookup_uccode

like lookup_catcode but targets Uccode and its table

lookup_value

lookup_vecdeque

like lookup_value but only recognizes Stored::VecDequeStored

maybe_note_label

Use to note a discovered label to support label-derived refererence numbers Can by used by \label, among others. Note we only record the label if it hasn’t already been peeked, and consumed.

merge_font

mouth_is_open

new_counter

Defines a new counter named $ctr. If within is defined, ctr will be reset whenever within is incremented.

new_local_box_list

note_begin

note_end

note_progress

note_progress_detailed

note_status

open_mouth

parse_parameters

If calling at compile-time, pass None for state, to avoid initialization.

parse_prototype

If calling at compile-time, pass None for state, to avoid initialization.

pin

Assign a string into the arena, returning a unique symbol associated with it

pin_char

pin_static

Assign a static str into the arena, returning a unique symbol associated with it

pop_box_list

pop_frame

Ends the current level of grouping. Note that this is lower level than \egroup;

pop_stack_frame

Removes the last/current stack frame, ending a TeX group

pop_value

pops the last value in a named Stored::VecDequeStored queue, if any

preload_font_map

process_options

progress_step

push_box_list

push_frame

Starts a new level of grouping. Note that this is lower level than C<\bgroup>;

push_pending_resource

push_stack_frame

Adds a new stack frame for a TeX group.

push_tokens

Pushes Tokens into a Stored::Tokens value when defined, or assigns when new.

push_value

manage a (global) list of values

raw_tex

a convenience function for including chunks of raw TeX (or LaTeX) code It is useful for copying portions of the normal implementation that can be handled simply using macros and primitives.

read_arg

read_balanced

read_balanced approximates TeX’s scan_toks (but doesn’t parse \def parameter lists) and only optionally requires the openning “{”. It may return comments in the token lists. it optionally (do_expand) expands while reading, but deferring \the and related. The is_macrodef flag affects whether # parameters are “packed” for macro bodies. If require_open is true, the opening T_BEGIN has not yet been read, and is required.

read_dimension

read_float

====================================================================== Float, a floating point number. Similar to factor, but does NOT accept comma! This is NOT part of TeX, but is convenient.

read_glue

read_keyword

Match the input against a set of keywords; Similar to readMatch, but the keywords are strings, and Case and catcodes are ignored; additionally, leading spaces are skipped. AND, macros are expanded.

read_match

Match the input against one of the Token or Tokens in @choices; return the matching one or undef.

read_mu_dimension

read_mu_glue

read_mu_unit

read_next_conditional

reads and discards tokens, until it encounters a conditional, if any

read_non_space

Read a single non-space token

read_normal_integer

read_number

read_optional

Read and return a LaTeX optional argument; returns default if there is no ‘[’, otherwise the contents of the array. Note that this returns an empty array if [] is present, i.e. [contents] in TeX will lead to Tokens(contents), otherwise returns None

read_raw_line

Read the next raw line (string); primarily to read from the Mouth, but keep any unread input!

read_register_value

read_rubber

read_token

read_tokens_value

Apparent behaviour of a token value (ie \toks#=<arg>)

read_until

Return a (balanced) sequence tokens until a match against one of the Tokens in @delims. Note that Braces on input hides the contents from matching, so this assumes there wont be braces in $delim! But, see readUntilBrace for that case.

read_until_brace

reads until it encounters a Catcode::BEGIN token

read_until_token

Convenience method wrapping around read_until TODO: This seems to be the wrong Rust type interface, we need to rework…

read_value

read_x_non_space

Read a single expanded, non-space, token

read_x_token

Read the next non-expandable token (expanding tokens until there’s a non-expandable one).

reading_from_mouth

Do something, while reading tokens from a specific Mouth. This reads ONLY from that mouth (or any mouth openned by code in that source), and the mouth should end up empty afterwards, and only be closed here.

ref_step_counter

Analog of \refstepcounter, steps the counter and returns a hash containing the keys refnum and id. This makes it suitable for use in a properties option to constructors. The id is generated in parallel with the reference number to assist debugging.

ref_step_id

For UN-numbered units. Like RefStepCounter, but only steps the “uncounter”, and returns only the id; This is useful for unnumbered cases of objects that normally get both a refnum and id.

ref_step_item_counter

Create id, and tags for an itemize type \item

regurgitate

steal the previously digested boxes from the current level.

remove_value

inline lookup_value after which globally assign an empty Tokens() to undo

remove_vecdeque

require_package

This (and FindFile) needs to evolve a bit to support reading raw .sty (.def, etc) files from the standard texmf directories. Maybe even use kpsewhich itself (INSTEAD of pathname_find ???) Another potentially useful option might be that if we are reading a raw file, perhaps it should just get digested immediately, since it shouldn’t contribute any boxes.

require_resource

rescue_caption_counters

Copies the current id, tags, and inlist counter values into whatsit properties

reset_counter

Resets the counter ctr to zero.

reset_pending_resources

resource_type

roman_aux

auxiliary helper for roman

search_paths_push_front

select_relaxng_schema

Selects the RelaxNG schema defining the XML output language

set_align_group_count

set_bindings_dispatch

set_build_template

set_dual_branch

sets the (localized) flag for “dual branch”

set_extra_bindings_dispatch

set_ifframe

sets a (originally Perl-local) IfFrame that needs to be manually expired.

set_indirect_model

set_input_encoding

set_mode

This sets the mode without doing any grouping (NOR does it stack the modes!!) Useful for environments, where the group has already been established. (presumably, in the long run, modes & groups should be much less coupled)

set_nomathparse_flag

set_prefix

Set one of the definition prefixes global, etc (only global matters!)

set_state

set_stomach

shift_value

skip_one_space

skip_spaces

state_is_unlocked

step_counter

Analog of \stepcounter, steps the counter and returns the expansion of \the$ctr Usually you should use ref_step_counter(ctr) instead.

take_build_template

take_pending_resources

to_string

unread

Push the tokens back into the input stream to be re-read.

unread_mut

Variant of unread, but drains the contents of tokens without taking ownership.

unread_one

Unreads a single Token to the start of the token stream

unread_vec

Unreads a Vec<Token> to the start of the token stream

unshift_value

use_main_state

use_std_state

use_sty_state

with

with2

with3

with_current_build_template

with_font_info

with_many

with_mapping

manage a (global) hash of values

with_mapping_keys

with_mapping_sym

with_mouth_mut

Execute a function with a mutable reference to the current mouth

with_search_paths

with_stacked_values

with_tag_property

with_tag_property_mut

with_value

with_value_mut

with_vecdeque

x_equals

XEquals check for two token arguments

Type Aliases

AssignmentCount

Ledger for stacked assignments

BeforeDigestClosure

ConditionalClosure

ConstructionClosure

DigestedReversionClosure

DigestionClosure

ExpansionClosure

FontClosure

FontTestClosure

HashMap

Type alias for a hash map that uses the Fx hashing algorithm.

LigatureMatcher

PrimitiveClosure

PrimitiveFn

ReaderClosure

ReplacementClosure

Result

ReversionClosure

StashTable

The (table_name, key, value) contents of a stored table of assignments

SymStr

Table

There are tables for catcode: keys are char; Also, math:char =1 when char is active in math. mathcode, sfcode, lccode, uccode, delcode : are similar to catcode but store additional kinds codes per char (see TeX) value: keys are anything (typically a string, though) and value is the value associated with it meaning: The definition assocated with key, usually a control-sequence. stash & stash_active: support named scopes (see also activateScope & deactivateScope)